fixed god object
