+++ mux-compiler/src/main.rs
@@ -90,7 +90,11 @@ fn find_clang_command() -> Option<String> {
 
     let linked_major = env!("MUX_LLVM_MAJOR");
     let versioned = format!("clang-{}", linked_major);
-    let candidates = [versioned.as_str(), "clang", "clang-17"];
+    let candidates: &[&str] = if linked_major != "17" {
+        &[versioned.as_str(), "clang", "clang-17"]
+    } else {
+        &[versioned.as_str(), "clang"]
+    };
     for candidate in candidates {
         let output = match Command::new(candidate).arg("--version").output() {
             Ok(output) => output,
