+++ mux-compiler/src/semantics/mod.rs
@@ -1,7 +1,13 @@
 // Module declarations
 
+pub mod declarations;
 pub mod error;
+pub mod expressions;
 pub mod format;
+pub mod free_vars;
+pub mod imports;
+pub mod patterns;
+pub mod statements;
 pub mod std_registry;
 pub mod stdlib;
 pub mod symbol_table;
@@ -17,13 +23,12 @@ pub use unifier::Unifier;
 
 // Internal imports
 use crate::ast::{
-    AstNode, BinaryOp, EnumVariant, ExpressionKind, ExpressionNode, Field, FunctionNode,
-    ImportSpec, LiteralNode, Param, PatternNode, PrimitiveType, Spanned, StatementKind,
-    StatementNode, TraitBound, TraitRef, TypeKind, TypeNode, UnaryOp,
+    AstNode, BinaryOp, ExpressionKind, ExpressionNode, Param, PrimitiveType, StatementKind,
+    StatementNode, TraitBound, TypeKind, TypeNode, UnaryOp,
 };
 use crate::diagnostic::Files;
 use crate::lexer::Span;
-use crate::semantics::std_registry::{StdModuleKind, std_module_registry};
+use crate::semantics::std_registry::std_module_registry;
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
 use std::rc::Rc;
@@ -34,7 +39,7 @@ type ResolvedInterface = (Vec<Type>, HashMap<String, MethodSig>);
 type ClassFieldInfo = (Type, bool);
 
 pub struct SemanticAnalyzer {
-    symbol_table: SymbolTable,
+    pub(super) symbol_table: SymbolTable,
     current_bounds: std::collections::HashMap<String, GenericBounds>,
     errors: Vec<SemanticError>,
     is_in_static_method: bool,
@@ -44,8 +49,8 @@ pub struct SemanticAnalyzer {
         std::collections::HashMap<String, std::collections::HashMap<String, Symbol>>,
     pub all_module_asts: std::collections::HashMap<String, Vec<AstNode>>,
     pub module_dependencies: Vec<String>,
-    required_runtime_features: HashSet<String>,
-    current_file: Option<std::path::PathBuf>, // Track current file for relative imports
+    pub(super) required_runtime_features: HashSet<String>,
+    pub(super) current_file: Option<std::path::PathBuf>, // Track current file for relative imports
     pub lambda_captures: std::collections::HashMap<Span, Vec<(String, Type)>>, // Track captured variables for each lambda
     pub current_return_type: Option<Type>, // Track current function/lambda return type
     pub current_class_type_params: Option<Vec<(String, GenericBounds)>>, // Track class-level type params with bounds for method analysis
@@ -60,7 +65,7 @@ impl Default for SemanticAnalyzer {
 
 impl SemanticAnalyzer {
     // Helper function to sanitize module paths for use in identifiers
-    fn sanitize_module_path(module_path: &str) -> String {
+    pub(super) fn sanitize_module_path(module_path: &str) -> String {
         module_path.replace(['.', '/'], "_")
     }
 
@@ -94,7 +99,7 @@ impl SemanticAnalyzer {
         }
     }
 
-    fn make_symbol(kind: SymbolKind, span: Span, type_: Option<Type>) -> Symbol {
+    pub(super) fn make_symbol(kind: SymbolKind, span: Span, type_: Option<Type>) -> Symbol {
         Symbol {
             kind,
             span,
@@ -110,7 +115,7 @@ impl SemanticAnalyzer {
         }
     }
 
-    fn make_module_symbol(&self, module_name: &str, span: Span) -> Symbol {
+    pub(super) fn make_module_symbol(&self, module_name: &str, span: Span) -> Symbol {
         Self::make_symbol(
             SymbolKind::Import,
             span,
@@ -172,7 +177,7 @@ impl SemanticAnalyzer {
 
     /// Inject nested stdlib children for a given parent stdlib module into
     /// `self.imported_symbols`. `parent_module` is the short name (e.g. "net").
-    fn inject_nested_stdlib_children(&mut self, parent_module: &str, span: Span) {
+    pub(super) fn inject_nested_stdlib_children(&mut self, parent_module: &str, span: Span) {
         let map = Self::stdlib_nested_modules_map();
         if let Some(children) = map.get(parent_module) {
             for child in children {
@@ -251,7 +256,7 @@ impl SemanticAnalyzer {
         features
     }
 
-    fn track_runtime_features_for_std_module_name(&mut self, module_name: &str) {
+    pub(super) fn track_runtime_features_for_std_module_name(&mut self, module_name: &str) {
         let registry = std_module_registry();
         let full_name = if module_name.starts_with("std.") {
             module_name.to_string()
@@ -521,7 +526,7 @@ impl SemanticAnalyzer {
         infer_missing_type_params_from_bounds(&func_node.type_params, substitutions);
     }
 
-    fn get_builtin_sig(&self, name: &str) -> Option<&BuiltInSig> {
+    pub(super) fn get_builtin_sig(&self, name: &str) -> Option<&BuiltInSig> {
         // Use the canonical BUILT_IN_FUNCTIONS from the stdlib module
         crate::semantics::stdlib::BUILT_IN_FUNCTIONS.get(name)
     }
@@ -2825,4156 +2830,6 @@ impl SemanticAnalyzer {
         } else {
             None
         }
-    }
-
-    // first pass, collect hoistable declarations like functions and classes.
-    fn collect_hoistable_declarations(&mut self, ast: &[AstNode]) -> Result<(), SemanticError> {
-        for node in ast {
-            match node {
-                AstNode::Function(func) => {
-                    self.collect_function_symbol(func)?;
-                }
-                AstNode::Class {
-                    name,
-                    traits,
-                    fields,
-                    methods,
-                    type_params,
-                    ..
-                } => {
-                    self.collect_class_symbol(
-                        name,
-                        traits,
-                        fields,
-                        methods,
-                        type_params,
-                        node.span(),
-                    )?;
-                }
-                AstNode::Enum { name, variants, .. } => {
-                    self.collect_enum_symbol(name, variants, node.span());
-                }
-                AstNode::Interface {
-                    name,
-                    type_params,
-                    fields,
-                    methods,
-                    ..
-                } => {
-                    self.collect_interface_symbol(name, type_params, fields, methods, node.span());
-                }
-                _ => {}
-            }
-        }
-        Ok(())
-    }
-
-    fn collect_function_symbol(&mut self, func: &FunctionNode) -> Result<(), SemanticError> {
-        if func.is_common {
-            return Err(SemanticError::with_help(
-                "Common methods are only allowed in classes",
-                func.span,
-                "The 'common' modifier creates a static method. Move this function inside a class definition, or remove the 'common' keyword.",
-            ));
-        }
-        let param_types = func
-            .params
-            .iter()
-            .map(|p| self.resolve_type(&p.type_))
-            .collect::<Result<Vec<_>, _>>()?;
-        let return_type = self.resolve_type(&func.return_type)?;
-        let mut func_type = Type::Function {
-            params: param_types,
-            returns: Box::new(return_type),
-            default_count: 0,
-        };
-
-        for (type_param_name, _) in &func.type_params {
-            let var = Type::Variable(type_param_name.clone());
-            func_type = self.substitute_type_param(&func_type, type_param_name, &var);
-        }
-
-        let default_count = func
-            .params
-            .iter()
-            .filter(|p| p.default_value.is_some())
-            .count();
-
-        if let Err(e) = self.symbol_table.add_symbol(
-            &func.name,
-            self.make_function_symbol(func.span, func_type, &func.type_params, default_count),
-        ) {
-            self.errors.push(e);
-        }
-        Ok(())
-    }
-
-    fn collect_class_symbol(
-        &mut self,
-        name: &str,
-        traits: &[TraitRef],
-        fields: &[Field],
-        methods: &[FunctionNode],
-        type_params: &[(String, Vec<TraitBound>)],
-        span: &Span,
-    ) -> Result<(), SemanticError> {
-        let implemented_interfaces = self.resolve_implemented_interfaces(traits, span)?;
-        let type_param_bounds: Vec<(String, Vec<String>)> = type_params
-            .iter()
-            .map(|(p, b)| (p.clone(), b.iter().map(|tb| tb.name.clone()).collect()))
-            .collect();
-
-        for (param_name, _) in type_params {
-            let _ = self.symbol_table.add_symbol(
-                param_name,
-                Self::make_symbol(
-                    SymbolKind::Type,
-                    *span,
-                    Some(Type::Generic(param_name.clone())),
-                ),
-            );
-        }
-
-        let (fields_map, _) = self.collect_class_fields(name, fields, span)?;
-        let methods_map = self.collect_class_methods(methods, name, type_params)?;
-
-        self.validate_interface_implementations(
-            name,
-            &implemented_interfaces,
-            &fields_map,
-            &methods_map,
-            span,
-        );
-
-        if let Err(e) = self.symbol_table.add_symbol(
-            name,
-            Symbol {
-                kind: SymbolKind::Class,
-                span: *span,
-                type_: Some(Type::Named(name.to_string(), vec![])),
-                interfaces: implemented_interfaces,
-                methods: methods_map,
-                fields: fields_map,
-                type_params: type_param_bounds,
-                original_name: None,
-                llvm_name: None,
-                default_param_count: 0,
-                variants: None,
-            },
-        ) {
-            self.errors.push(e);
-        }
-        Ok(())
-    }
-
-    fn resolve_implemented_interfaces(
-        &self,
-        traits: &[TraitRef],
-        _span: &Span,
-    ) -> Result<HashMap<String, ResolvedInterface>, SemanticError> {
-        let mut implemented_interfaces = std::collections::HashMap::new();
-        for trait_ref in traits {
-            if let Some(interface_symbol) = self.symbol_table.lookup(&trait_ref.name)
-                && let Some((_, interface_methods)) =
-                    interface_symbol.interfaces.get(&trait_ref.name)
-            {
-                let resolved_args = trait_ref
-                    .type_args
-                    .iter()
-                    .map(|arg| self.resolve_type(arg))
-                    .collect::<Result<Vec<_>, _>>()?;
-                let interface_type_params = &interface_symbol.type_params;
-                let mut substituted_methods = std::collections::HashMap::new();
-                for (method_name, method_sig) in interface_methods {
-                    let sub_params = method_sig
-                        .params
-                        .iter()
-                        .map(|p| {
-                            self.substitute_type_params(p, interface_type_params, &resolved_args)
-                        })
-                        .collect();
-                    let sub_return = self.substitute_type_params(
-                        &method_sig.return_type,
-                        interface_type_params,
-                        &resolved_args,
-                    );
-                    substituted_methods.insert(
-                        method_name.clone(),
-                        MethodSig {
-                            params: sub_params,
-                            return_type: sub_return,
-                            is_static: method_sig.is_static,
-                        },
-                    );
-                }
-                implemented_interfaces
-                    .insert(trait_ref.name.clone(), (resolved_args, substituted_methods));
-            }
-        }
-        Ok(implemented_interfaces)
-    }
-
-    fn collect_class_fields(
-        &mut self,
-        name: &str,
-        fields: &[Field],
-        _span: &Span,
-    ) -> Result<(HashMap<String, ClassFieldInfo>, Vec<Type>), SemanticError> {
-        let mut field_types = Vec::new();
-        let mut fields_map = std::collections::HashMap::new();
-        for field in fields {
-            if fields_map.contains_key(&field.name) {
-                return Err(SemanticError::with_help(
-                    format!("Duplicate field '{}' in class '{}'", field.name, name),
-                    field.type_.span,
-                    "Each field name must be unique within a class. Rename or remove the duplicate field.",
-                ));
-            }
-            match self.resolve_type(&field.type_) {
-                Ok(t) => {
-                    field_types.push(t.clone());
-                    fields_map.insert(field.name.clone(), (t, field.is_const));
-                }
-                Err(e) => self.errors.push(e),
-            }
-        }
-        Ok((fields_map, field_types))
-    }
-
-    fn collect_class_methods(
-        &mut self,
-        methods: &[FunctionNode],
-        class_name: &str,
-        type_params: &[(String, Vec<TraitBound>)],
-    ) -> Result<HashMap<String, MethodSig>, SemanticError> {
-        let mut methods_map = HashMap::new();
-        for method in methods {
-            let mut param_types = Vec::new();
-            for p in &method.params {
-                match self.resolve_type(&p.type_) {
-                    Ok(t) => param_types.push(t),
-                    Err(e) => self.errors.push(e),
-                }
-            }
-            let ret = match self.resolve_type(&method.return_type) {
-                Ok(t) => t,
-                Err(e) => {
-                    self.errors.push(e);
-                    continue;
-                }
-            };
-            let method_sig = MethodSig {
-                params: param_types,
-                return_type: ret,
-                is_static: method.is_common,
-            };
-            methods_map.insert(method.name.clone(), method_sig);
-        }
-
-        let type_args: Vec<Type> = type_params
-            .iter()
-            .map(|(p, _)| Type::Variable(p.clone()))
-            .collect();
-        let new_sig = MethodSig {
-            params: vec![],
-            return_type: Type::Named(class_name.to_string(), type_args),
-            is_static: true,
-        };
-        methods_map.insert("new".to_string(), new_sig);
-        Ok(methods_map)
-    }
-
-    fn validate_interface_implementations(
-        &mut self,
-        class_name: &str,
-        implemented_interfaces: &HashMap<String, ResolvedInterface>,
-        fields_map: &HashMap<String, ClassFieldInfo>,
-        methods_map: &HashMap<String, MethodSig>,
-        span: &Span,
-    ) {
-        for (interface_name, (_, interface_methods)) in implemented_interfaces {
-            self.validate_interface_methods(
-                class_name,
-                interface_name,
-                interface_methods,
-                methods_map,
-                *span,
-            );
-            self.validate_interface_fields(class_name, interface_name, fields_map, *span);
-        }
-    }
-
-    fn validate_interface_methods(
-        &mut self,
-        class_name: &str,
-        interface_name: &str,
-        interface_methods: &HashMap<String, MethodSig>,
-        methods_map: &HashMap<String, MethodSig>,
-        span: Span,
-    ) {
-        for (method_name, interface_sig) in interface_methods {
-            if let Some(class_sig) = methods_map.get(method_name) {
-                if let Err(e) = self.check_method_compatibility(interface_sig, class_sig, span) {
-                    self.errors.push(e);
-                }
-            } else {
-                self.errors.push(SemanticError::with_help(
-                    format!(
-                        "Class '{}' does not implement method '{}' required by interface '{}'",
-                        class_name, method_name, interface_name
-                    ),
-                    span,
-                    format!("Add a method '{}' to class '{}' with the signature required by interface '{}'", method_name, class_name, interface_name),
-                ));
-            }
-        }
-    }
-
-    fn validate_interface_fields(
-        &mut self,
-        class_name: &str,
-        interface_name: &str,
-        fields_map: &HashMap<String, ClassFieldInfo>,
-        span: Span,
-    ) {
-        if let Some(interface_symbol) = self.symbol_table.lookup(interface_name) {
-            for (field_name, (interface_field_type, interface_is_const)) in &interface_symbol.fields
-            {
-                if let Some((class_field_type, class_is_const)) = fields_map.get(field_name) {
-                    self.validate_field_type(
-                        class_name,
-                        interface_name,
-                        field_name,
-                        interface_field_type,
-                        class_field_type,
-                        span,
-                    );
-                    self.validate_field_const(
-                        class_name,
-                        interface_name,
-                        field_name,
-                        *interface_is_const,
-                        *class_is_const,
-                        span,
-                    );
-                } else {
-                    self.errors.push(SemanticError::with_help(
-                        format!(
-                            "Class '{}' is missing required field '{}' from interface '{}'",
-                            class_name, field_name, interface_name
-                        ),
-                        span,
-                        format!(
-                            "Add field '{}: {}' to class '{}'",
-                            field_name,
-                            format_type(interface_field_type),
-                            class_name
-                        ),
-                    ));
-                }
-            }
-        }
-    }
-
-    fn validate_field_type(
-        &mut self,
-        class_name: &str,
-        interface_name: &str,
-        field_name: &str,
-        interface_field_type: &Type,
-        class_field_type: &Type,
-        span: Span,
-    ) {
-        if !self.types_compatible(class_field_type, interface_field_type) {
-            self.errors.push(SemanticError::with_help(
-                format!(
-                    "Field '{}' type mismatch in class '{}': class has {}, interface '{}' requires {}",
-                    field_name,
-                    class_name,
-                    format_type(class_field_type),
-                    interface_name,
-                    format_type(interface_field_type)
-                ),
-                span,
-                format!("Change the type of field '{}' to {} to match interface '{}'", field_name, format_type(interface_field_type), interface_name),
-            ));
-        }
-    }
-
-    fn validate_field_const(
-        &mut self,
-        class_name: &str,
-        interface_name: &str,
-        field_name: &str,
-        interface_is_const: bool,
-        class_is_const: bool,
-        span: Span,
-    ) {
-        if interface_is_const && !class_is_const {
-            self.errors.push(SemanticError::with_help(
-                format!(
-                    "Field '{}' must be const in class '{}' to implement interface '{}'",
-                    field_name, class_name, interface_name
-                ),
-                span,
-                format!(
-                    "Add the 'const' modifier to field '{}' in class '{}'",
-                    field_name, class_name
-                ),
-            ));
-        }
-    }
-
-    fn collect_enum_symbol(&mut self, name: &str, variants: &[EnumVariant], span: &Span) {
-        let mut methods = std::collections::HashMap::new();
-        let mut variant_names = Vec::new();
-        for variant in variants {
-            variant_names.push(variant.name.clone());
-            let params = variant
-                .data
-                .clone()
-                .unwrap_or_default()
-                .into_iter()
-                .map(|(_, t)| self.resolve_type(&t).unwrap_or(Type::Void))
-                .collect();
-            let return_type = Type::Named(name.to_string(), vec![]);
-            methods.insert(
-                variant.name.clone(),
-                MethodSig {
-                    params,
-                    return_type,
-                    is_static: true,
-                },
-            );
-        }
-        if let Err(e) = self.symbol_table.add_symbol(
-            name,
-            Symbol {
-                kind: SymbolKind::Enum,
-                span: *span,
-                type_: Some(Type::Named(name.to_string(), vec![])),
-                interfaces: std::collections::HashMap::new(),
-                methods,
-                fields: std::collections::HashMap::new(),
-                type_params: Vec::new(),
-                original_name: None,
-                llvm_name: None,
-                default_param_count: 0,
-                variants: Some(variant_names),
-            },
-        ) {
-            self.errors.push(e);
-        }
-    }
-
-    fn collect_interface_symbol(
-        &mut self,
-        name: &str,
-        type_params: &[(String, Vec<TraitBound>)],
-        fields: &[Field],
-        methods: &[FunctionNode],
-        span: &Span,
-    ) {
-        let mut interface_methods = std::collections::HashMap::new();
-        for method in methods {
-            let param_types = method
-                .params
-                .iter()
-                .map(|p| self.resolve_type(&p.type_))
-                .collect::<Result<Vec<_>, _>>();
-            if let Ok(param_types) = param_types {
-                let return_type = self.resolve_type(&method.return_type);
-                if let Ok(return_type) = return_type {
-                    let method_sig = MethodSig {
-                        params: param_types,
-                        return_type,
-                        is_static: false,
-                    };
-                    interface_methods.insert(method.name.clone(), method_sig);
-                }
-            }
-        }
-
-        let mut interface_fields = std::collections::HashMap::new();
-        for field in fields {
-            if let Ok(field_type) = self.resolve_type(&field.type_) {
-                if let Some(default_expr) = &field.default_value
-                    && let Ok(default_type) = self.infer_literal_type(default_expr)
-                    && !self.types_compatible(&default_type, &field_type)
-                {
-                    self.errors.push(SemanticError::with_help(
-                        format!(
-                            "Default value type mismatch for field '{}': expected {}, got {}",
-                            field.name,
-                            format_type(&field_type),
-                            format_type(&default_type)
-                        ),
-                        default_expr.span,
-                        format!(
-                            "The default value must match the field's declared type of {}",
-                            format_type(&field_type)
-                        ),
-                    ));
-                }
-                interface_fields.insert(field.name.clone(), (field_type, field.is_const));
-            }
-        }
-
-        let mut interfaces_map = std::collections::HashMap::new();
-        interfaces_map.insert(name.to_string(), (vec![], interface_methods));
-        if let Err(e) = self.symbol_table.add_symbol(
-            name,
-            Symbol {
-                kind: SymbolKind::Interface,
-                span: *span,
-                type_: None,
-                interfaces: interfaces_map,
-                methods: std::collections::HashMap::new(),
-                fields: interface_fields,
-                type_params: type_params
-                    .iter()
-                    .map(|(p, b)| (p.clone(), b.iter().map(|tb| tb.name.clone()).collect()))
-                    .collect::<Vec<(String, Vec<String>)>>(),
-                original_name: None,
-                llvm_name: None,
-                default_param_count: 0,
-                variants: None,
-            },
-        ) {
-            self.errors.push(e);
-        }
-    }
-
-    // second pass, analyze all nodes with full symbol information.
-    fn analyze_nodes(&mut self, nodes: &[AstNode], mut files: Option<&mut Files>) {
-        for node in nodes {
-            if let Err(e) = self.analyze_node(node, files.as_deref_mut()) {
-                self.errors.push(e);
-            }
-        }
-    }
-
-    fn analyze_node(
-        &mut self,
-        node: &AstNode,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        match node {
-            AstNode::Function(func) => {
-                if func.is_common {
-                    return Err(SemanticError::with_help(
-                        "Common methods are only allowed inside class definitions",
-                        func.span,
-                        "The 'common' modifier creates static methods on a class. Remove 'common' for standalone functions.",
-                    ));
-                }
-
-                // Check that main() returns void
-                if func.name == "main" {
-                    let return_type = self.resolve_type(&func.return_type)?;
-                    if !matches!(return_type, Type::Void) {
-                        return Err(SemanticError::with_help(
-                            format!(
-                                "Function 'main' must return void, not '{}'",
-                                format_type(&return_type)
-                            ),
-                            func.return_type.span,
-                            "The entry point 'main' function must be declared as 'returns void'. Change the return type to void.",
-                        ));
-                    }
-                }
-
-                self.analyze_function(func, None)
-            }
-            AstNode::Class {
-                name,
-                fields,
-                methods,
-                type_params,
-                ..
-            } => {
-                let type_param_bounds = self.resolve_type_param_bounds(type_params)?;
-                self.analyze_class(name, fields, methods, &type_param_bounds)
-            }
-            AstNode::Enum { .. } => Ok(()), // enums don't need further analysis.
-            AstNode::Interface { .. } => Ok(()), // interfaces don't need further analysis.
-            AstNode::Statement(stmt) => self.analyze_statement(stmt, files),
-        }
-    }
-
-    fn analyze_function(
-        &mut self,
-        func: &FunctionNode,
-        self_type: Option<Type>,
-    ) -> Result<(), SemanticError> {
-        let was_static = self.is_in_static_method;
-        let old_self_type = self.current_self_type.clone();
-        let old_return_type = self.current_return_type.clone();
-        self.is_in_static_method = func.is_common;
-        self.current_self_type = self_type.clone();
-
-        // Set return type context
-        let return_type = self.resolve_type(&func.return_type)?;
-        self.current_return_type = Some(return_type.clone());
-
-        // set current bounds
-        self.current_bounds.clear();
-        for (param, bounds) in self.resolve_type_param_bounds(&func.type_params)? {
-            self.current_bounds.insert(param, bounds);
-        }
-
-        // Also add class-level type params (for methods in generic classes)
-        if let Some(class_type_params) = &self.current_class_type_params {
-            for (param, bounds) in class_type_params {
-                // Don't override function-level type params
-                if !self.current_bounds.contains_key(param) {
-                    self.current_bounds.insert(param.clone(), bounds.clone());
-                }
-            }
-        }
-
-        // create new scope for function parameters and body.
-        self.symbol_table.push_scope()?;
-
-        // add generic type parameters to symbol table
-        for (param_name, _) in &func.type_params {
-            self.symbol_table.add_symbol(
-                param_name,
-                Symbol {
-                    kind: SymbolKind::Type,
-                    span: func.span, // Use function span since we don't have param spans
-                    type_: Some(Type::Generic(param_name.clone())),
-                    interfaces: std::collections::HashMap::new(),
-                    methods: std::collections::HashMap::new(),
-                    fields: std::collections::HashMap::new(),
-                    type_params: Vec::new(),
-                    original_name: None,
-                    llvm_name: None,
-                    default_param_count: 0,
-                    variants: None,
-                },
-            )?;
-        }
-
-        // add self if provided
-        if let Some(self_type) = self_type {
-            self.symbol_table.add_symbol(
-                "self",
-                Self::make_symbol(SymbolKind::Variable, func.span, Some(self_type)),
-            )?;
-        }
-
-        for param in &func.params {
-            let param_type = self.resolve_type(&param.type_)?;
-            self.symbol_table.add_symbol(
-                &param.name,
-                Self::make_symbol(SymbolKind::Variable, param.type_.span, Some(param_type)),
-            )?;
-        }
-
-        // analyze function body with new scope.
-        self.analyze_block(&func.body, None)?;
-
-        // clean up function scope.
-        self.symbol_table.pop_scope()?;
-        self.current_bounds.clear();
-        self.is_in_static_method = was_static;
-        self.current_self_type = old_self_type;
-        self.current_return_type = old_return_type;
-
-        if func.body.is_empty() || !self.all_paths_return(&func.body) {
-            let (msg, help): (String, String) = if matches!(return_type, Type::Void) {
-                (
-                    "Function must end with an explicit 'return' statement on all code paths"
-                        .to_string(),
-                    "Add a 'return' statement at the end of every code path".to_string(),
-                )
-            } else {
-                (
-                    format!(
-                        "Function must return a value of type '{}' on all code paths",
-                        format_type(&return_type)
-                    ),
-                    "Add a return statement at the end of every branch (if/else, match, etc.)"
-                        .to_string(),
-                )
-            };
-            return Err(SemanticError::with_help(msg, func.span, help));
-        }
-
-        Ok(())
-    }
-
-    #[allow(clippy::only_used_in_recursion)]
-    /// Returns true if the given expression is a statically-known infinite loop
-    /// condition, currently the literal `true`. This lets `while true { ... }`
-    /// be recognized as always returning when its body always returns.
-    fn is_infinite_loop_condition(cond: &ExpressionNode) -> bool {
-        matches!(
-            &cond.kind,
-            ExpressionKind::Literal(LiteralNode::Boolean(true))
-        )
-    }
-
-    #[allow(clippy::only_used_in_recursion)]
-    fn all_paths_return(&self, stmts: &[StatementNode]) -> bool {
-        stmts.iter().any(|stmt| self.statement_returns(stmt))
-    }
-
-    /// Returns true if the given statement terminates this code path with a
-    /// return. Used by `all_paths_return` to keep the recursive function
-    /// under SonarQube's cognitive complexity threshold.
-    #[allow(clippy::only_used_in_recursion)]
-    fn statement_returns(&self, stmt: &StatementNode) -> bool {
-        match &stmt.kind {
-            StatementKind::Return(_) => true,
-            StatementKind::If {
-                then_block,
-                else_block,
-                ..
-            } => {
-                let else_returns = else_block
-                    .as_ref()
-                    .is_some_and(|b| self.all_paths_return(b));
-                self.all_paths_return(then_block) && else_returns
-            }
-            StatementKind::Block(block_stmts) => self.all_paths_return(block_stmts),
-            StatementKind::While { cond, body: _ } => Self::is_infinite_loop_condition(cond),
-            StatementKind::Match { arms, .. } => {
-                arms.iter().all(|arm| self.all_paths_return(&arm.body))
-            }
-            _ => false,
-        }
-    }
-
-    fn analyze_class(
-        &mut self,
-        name: &str,
-        _fields: &[Field],
-        methods: &[FunctionNode],
-        type_params: &[(String, GenericBounds)],
-    ) -> Result<(), SemanticError> {
-        // Methods were already added to the class symbol during first pass (collect_hoistable_declarations)
-        // Here we just need to analyze method bodies with proper self type
-
-        // Create self type for this class
-        let self_type = if type_params.is_empty() {
-            Type::Named(name.to_string(), vec![])
-        } else {
-            // For generic classes, use type variables for self
-            Type::Named(
-                name.to_string(),
-                type_params
-                    .iter()
-                    .map(|(param_name, _)| Type::Variable(param_name.clone()))
-                    .collect(),
-            )
-        };
-
-        // Set current class type params for method analysis
-        self.set_class_type_params(type_params.to_vec());
-
-        // Analyze each method body with proper self type
-        for method in methods {
-            // Static methods (common) don't have self
-            let method_self_type = if method.is_common {
-                None
-            } else {
-                Some(self_type.clone())
-            };
-
-            self.analyze_function(method, method_self_type)?;
-        }
-
-        // Clear class type params after analyzing class methods
-        self.clear_class_type_params();
-
-        Ok(())
-    }
-
-    fn analyze_block(
-        &mut self,
-        stmts: &[StatementNode],
-        mut files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        // first collect function declarations in this block.
-        for stmt in stmts {
-            if let StatementKind::Function(func) = &stmt.kind {
-                // resolve function type
-                let param_types = func
-                    .params
-                    .iter()
-                    .map(|p| self.resolve_type(&p.type_))
-                    .collect::<Result<Vec<_>, _>>()?;
-                let return_type = self.resolve_type(&func.return_type)?;
-                let mut func_type = Type::Function {
-                    params: param_types,
-                    returns: Box::new(return_type),
-                    default_count: 0,
-                };
-
-                // substitute type params with variables
-                for (type_param_name, _) in &func.type_params {
-                    let var = Type::Variable(type_param_name.clone());
-                    func_type = self.substitute_type_param(&func_type, type_param_name, &var);
-                }
-
-                // Count parameters with default values
-                let default_count = func
-                    .params
-                    .iter()
-                    .filter(|p| p.default_value.is_some())
-                    .count();
-
-                self.symbol_table.add_symbol(
-                    &func.name,
-                    self.make_function_symbol(
-                        stmt.span,
-                        func_type,
-                        &func.type_params,
-                        default_count,
-                    ),
-                )?;
-            }
-        }
-
-        // then analyze all statements in order.
-        for stmt in stmts {
-            self.analyze_statement(stmt, files.as_deref_mut())?;
-        }
-
-        Ok(())
-    }
-
-    fn analyze_auto_decl_statement(
-        &mut self,
-        name: &str,
-        expr: &ExpressionNode,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(expr)?;
-        let expr_type = self.get_expression_type(expr)?;
-        if matches!(
-            expr_type,
-            Type::EmptyList | Type::EmptyMap | Type::EmptySet | Type::EmptySetOrMap
-        ) {
-            let collection_type = match expr_type {
-                Type::EmptyList => "list",
-                Type::EmptyMap => "map",
-                Type::EmptySet => "set",
-                Type::EmptySetOrMap => "set or map",
-                _ => unreachable!(),
-            };
-            return Err(SemanticError::with_help(
-                format!("Cannot infer type for empty {} literal", collection_type),
-                expr.span,
-                format!(
-                    "Use an explicit type annotation, e.g. {}<int> myVar = {{}}",
-                    collection_type
-                ),
-            ));
-        }
-        if Self::type_contains_empty_set_or_map(&expr_type) {
-            return Err(SemanticError::with_help(
-                "Cannot infer type for expression containing empty set/map literal".to_string(),
-                expr.span,
-                "Use explicit type annotations for all empty collection literals.",
-            ));
-        }
-        self.symbol_table.add_symbol(
-            name,
-            Self::make_symbol(SymbolKind::Variable, span, Some(expr_type)),
-        )?;
-        Ok(())
-    }
-
-    fn type_contains_empty_set_or_map(ty: &Type) -> bool {
-        match ty {
-            Type::EmptySetOrMap => true,
-            Type::Map(key_type, value_type) => {
-                Self::type_contains_empty_set_or_map(key_type)
-                    || Self::type_contains_empty_set_or_map(value_type)
-            }
-            Type::Set(elem_type) => Self::type_contains_empty_set_or_map(elem_type),
-            Type::List(elem_type) => Self::type_contains_empty_set_or_map(elem_type),
-            Type::Optional(inner) => Self::type_contains_empty_set_or_map(inner),
-            Type::Result(ok, err) => {
-                Self::type_contains_empty_set_or_map(ok)
-                    || Self::type_contains_empty_set_or_map(err)
-            }
-            Type::Tuple(left, right) => {
-                Self::type_contains_empty_set_or_map(left)
-                    || Self::type_contains_empty_set_or_map(right)
-            }
-            _ => false,
-        }
-    }
-
-    fn analyze_typed_decl_statement(
-        &mut self,
-        name: &str,
-        type_node: &TypeNode,
-        expr: &ExpressionNode,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let declared_type = self.resolve_type(type_node)?;
-        self.analyze_expression(expr)?;
-        self.resolve_empty_collection_types(&declared_type, expr)?;
-        let expr_type = self.get_expression_type(expr)?;
-        self.check_type_compatibility(&declared_type, &expr_type, expr.span)?;
-        self.symbol_table.add_symbol(
-            name,
-            Self::make_symbol(SymbolKind::Variable, span, Some(declared_type)),
-        )?;
-        Ok(())
-    }
-
-    fn resolve_empty_collection_types(
-        &mut self,
-        expected_type: &Type,
-        expr: &ExpressionNode,
-    ) -> Result<(), SemanticError> {
-        match &expr.kind {
-            ExpressionKind::SetOrMapLiteral(elements) if elements.is_empty() => {
-                self.resolve_empty_set_or_map(expected_type, expr.span);
-            }
-            ExpressionKind::MapLiteral { entries, .. } => {
-                self.resolve_map_literal_children(expected_type, entries)?;
-            }
-            ExpressionKind::SetOrMapLiteral(elements) => {
-                self.resolve_typed_collection_elements(expected_type, elements)?;
-            }
-            ExpressionKind::ListLiteral(elements) => {
-                self.resolve_typed_collection_elements(expected_type, elements)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn resolve_empty_set_or_map(&mut self, expected_type: &Type, span: Span) {
-        let resolved = match expected_type {
-            Type::Map(_, _) => Type::EmptyMap,
-            Type::Set(_) => Type::EmptySet,
-            _ => return,
-        };
-        self.expression_type_overrides.insert(span, resolved);
-    }
-
-    fn resolve_map_literal_children(
-        &mut self,
-        expected_type: &Type,
-        entries: &[(ExpressionNode, ExpressionNode)],
-    ) -> Result<(), SemanticError> {
-        if let Type::Map(_, value_type) = expected_type {
-            for (_, value_expr) in entries {
-                self.resolve_empty_collection_types(value_type, value_expr)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn resolve_typed_collection_elements(
-        &mut self,
-        expected_type: &Type,
-        elements: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        let elem_type = match expected_type {
-            Type::Set(elem) => elem,
-            Type::List(elem) => elem,
-            _ => return Ok(()),
-        };
-        for element in elements {
-            self.resolve_empty_collection_types(elem_type, element)?;
-        }
-        Ok(())
-    }
-
-    fn analyze_const_decl_statement(
-        &mut self,
-        name: &str,
-        type_node: &TypeNode,
-        expr: &ExpressionNode,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let declared_type = self.resolve_type(type_node)?;
-        self.analyze_expression(expr)?;
-        self.resolve_empty_collection_types(&declared_type, expr)?;
-        let expr_type = self.get_expression_type(expr)?;
-        self.check_type_compatibility(&declared_type, &expr_type, expr.span)?;
-        self.symbol_table.add_symbol(
-            name,
-            Self::make_symbol(SymbolKind::Constant, span, Some(declared_type)),
-        )?;
-        Ok(())
-    }
-
-    fn analyze_if_statement(
-        &mut self,
-        cond: &ExpressionNode,
-        then_block: &[StatementNode],
-        else_block: &Option<Vec<StatementNode>>,
-        mut files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(cond)?;
-
-        self.symbol_table.push_scope()?;
-        self.analyze_block(then_block, files.as_deref_mut())?;
-        self.symbol_table.pop_scope()?;
-
-        if let Some(else_block) = else_block {
-            self.symbol_table.push_scope()?;
-            self.analyze_block(else_block, files)?;
-            self.symbol_table.pop_scope()?;
-        }
-
-        Ok(())
-    }
-
-    fn analyze_for_statement(
-        &mut self,
-        var: &str,
-        iter: &ExpressionNode,
-        body: &[StatementNode],
-        span: Span,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        self.symbol_table.push_scope()?;
-        self.analyze_expression(iter)?;
-        let iter_type = self.get_expression_type(iter)?;
-        let var_type = match iter_type {
-            Type::List(elem_type) => *elem_type,
-            _ => {
-                return Err(SemanticError::with_help(
-                    format!("Cannot iterate over type {}", format_type(&iter_type)),
-                    iter.span,
-                    "The 'for' loop can only iterate over list types. Use .to_list() for sets. For maps, use .get_keys(), .get_values(), or .get_pairs() to get a list. For numeric ranges, use range(start, end).",
-                ));
-            }
-        };
-        self.symbol_table.add_symbol(
-            var,
-            Self::make_symbol(SymbolKind::Variable, span, Some(var_type)),
-        )?;
-        self.analyze_block(body, files)?;
-        self.symbol_table.pop_scope()?;
-        Ok(())
-    }
-
-    fn analyze_while_statement(
-        &mut self,
-        cond: &ExpressionNode,
-        body: &[StatementNode],
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(cond)?;
-        self.symbol_table.push_scope()?;
-        self.analyze_block(body, files)?;
-        self.symbol_table.pop_scope()?;
-        Ok(())
-    }
-
-    fn collect_match_arm_returns(
-        &mut self,
-        arm: &crate::ast::MatchArm,
-        expr_span: Span,
-        expecting_return: bool,
-        arm_return_types: &mut Vec<(Type, Span)>,
-    ) -> Result<(), SemanticError> {
-        if !expecting_return {
-            return Ok(());
-        }
-
-        let mut has_return = false;
-        for stmt in &arm.body {
-            if let StatementKind::Return(Some(ret_expr)) = &stmt.kind {
-                let ret_type = self.get_expression_type(ret_expr)?;
-                arm_return_types.push((ret_type, stmt.span));
-                has_return = true;
-                break;
-            }
-            if let StatementKind::Return(None) = &stmt.kind {
-                has_return = true;
-            }
-        }
-
-        if has_return {
-            return Ok(());
-        }
-
-        Err(SemanticError::with_help(
-            "Match arm must return a value in a function with non-void return type",
-            arm.body.first().map(|s| s.span).unwrap_or(expr_span),
-            "Add a return statement to this match arm",
-        ))
-    }
-
-    fn validate_match_arm_return_types(
-        &self,
-        arm_return_types: &[(Type, Span)],
-    ) -> Result<(), SemanticError> {
-        if arm_return_types.is_empty() {
-            return Ok(());
-        }
-
-        let (first_type, _) = &arm_return_types[0];
-        for (i, (arm_type, arm_span)) in arm_return_types.iter().enumerate().skip(1) {
-            if self
-                .check_type_compatibility(first_type, arm_type, *arm_span)
-                .is_err()
-            {
-                return Err(SemanticError::with_help(
-                    format!(
-                        "Match arms have incompatible return types: first arm returns '{}', but arm {} returns '{}'",
-                        format_type(first_type),
-                        i + 1,
-                        format_type(arm_type)
-                    ),
-                    *arm_span,
-                    "All match arms must return the same type when used as an expression or in a returning context",
-                ));
-            }
-        }
-
-        Ok(())
-    }
-
-    fn analyze_match_statement(
-        &mut self,
-        expr: &ExpressionNode,
-        arms: &[crate::ast::MatchArm],
-        mut files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(expr)?;
-        let expr_type = self.get_expression_type(expr)?;
-
-        let expecting_return = self.current_return_type.is_some()
-            && !matches!(self.current_return_type, Some(Type::Void));
-        let mut arm_return_types = Vec::new();
-
-        for arm in arms {
-            self.symbol_table.push_scope()?;
-            self.set_pattern_types(&arm.pattern, &expr_type, expr.span)?;
-            self.analyze_pattern(&arm.pattern)?;
-            if let Some(guard) = &arm.guard {
-                self.analyze_expression(guard)?;
-            }
-            self.analyze_block(&arm.body, files.as_deref_mut())?;
-            self.collect_match_arm_returns(
-                arm,
-                expr.span,
-                expecting_return,
-                &mut arm_return_types,
-            )?;
-            self.symbol_table.pop_scope()?;
-        }
-
-        self.validate_match_arm_return_types(&arm_return_types)?;
-        self.check_match_exhaustiveness(&expr_type, arms, expr.span)
-    }
-
-    fn analyze_return_with_value(&mut self, expr: &ExpressionNode) -> Result<(), SemanticError> {
-        if self.current_return_type.is_none() {
-            return Err(SemanticError::with_help(
-                "Cannot use 'return' outside of a function",
-                expr.span,
-                "'return' can only be used inside a function body",
-            ));
-        }
-
-        self.analyze_expression(expr)?;
-
-        if matches!(self.current_return_type, Some(Type::Void)) {
-            return Err(SemanticError::with_help(
-                "Cannot return a value from a void function",
-                expr.span,
-                "This function is declared as 'returns void'. Either remove the return value or change the function's return type.",
-            ));
-        }
-
-        if let Some(expected_type) = self.current_return_type.clone() {
-            let actual_type = self.get_expression_type(expr)?;
-            self.check_type_compatibility(&expected_type, &actual_type, expr.span)?;
-        }
-        Ok(())
-    }
-
-    fn analyze_return_without_value(&mut self, span: Span) -> Result<(), SemanticError> {
-        if self.current_return_type.is_none() {
-            return Err(SemanticError::with_help(
-                "Cannot use 'return' outside of a function",
-                span,
-                "'return' can only be used inside a function body",
-            ));
-        }
-
-        if !matches!(self.current_return_type, Some(Type::Void))
-            && let Some(expected_type) = &self.current_return_type
-        {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Missing return value; function expects '{}', but return has no value",
-                    format_type(expected_type)
-                ),
-                span,
-                format!(
-                    "Add a value after 'return' that matches the function's return type '{}'. Example: return some_value",
-                    format_type(expected_type)
-                ),
-            ));
-        }
-        Ok(())
-    }
-
-    fn analyze_import_statement(
-        &mut self,
-        module_path: &str,
-        spec: &ImportSpec,
-        span: Span,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        if module_path == "std" || module_path.starts_with("std.") {
-            self.handle_std_import(module_path, spec, span, files)?;
-            return Ok(());
-        }
-
-        let resolver = self
-            .module_resolver
-            .clone()
-            .ok_or_else(|| SemanticError::new("Module resolver not available", span))?;
-
-        let (has_file, has_directory) = resolver.borrow().check_module_path(module_path);
-
-        if has_file && has_directory {
-            return Err(SemanticError::with_help(
-                format!("Ambiguous import: '{}'", module_path),
-                span,
-                format!(
-                    "Both {}.mux and {}/ directory exist. Please remove one.",
-                    module_path.replace('.', "/"),
-                    module_path.replace('.', "/")
-                ),
-            ));
-        }
-
-        let files = files.ok_or_else(|| {
-            SemanticError::new(
-                "Files registry must be available for import processing",
-                span,
-            )
-        })?;
-
-        if has_directory {
-            self.handle_directory_import(module_path, spec, span, resolver, files)?;
-            return Ok(());
-        }
-
-        let module_nodes = resolver
-            .borrow_mut()
-            .resolve_import_path(module_path, self.current_file.as_deref(), files)
-            .map_err(|e| {
-                SemanticError::with_help(
-                    format!("Failed to import module '{}'", module_path),
-                    span,
-                    e.to_string(),
-                )
-            })?;
-
-        let mut module_analyzer = SemanticAnalyzer::new_for_module(resolver.clone());
-        module_analyzer.set_current_file(std::path::PathBuf::from(
-            module_path.replace('.', "/") + ".mux",
-        ));
-        let errors = module_analyzer.analyze(&module_nodes, Some(files));
-        if !errors.is_empty() {
-            let error_messages: Vec<String> = errors.iter().map(|e| e.message.clone()).collect();
-            return Err(SemanticError::with_help(
-                format!("Errors in imported module '{}'", module_path),
-                span,
-                format!(
-                    "Fix the following errors in '{}':\n  {}",
-                    module_path,
-                    error_messages.join("\n  ")
-                ),
-            ));
-        }
-
-        let module_symbols =
-            self.filter_module_export_symbols(&module_analyzer.symbol_table.all_symbols);
-        self.required_runtime_features
-            .extend(module_analyzer.required_runtime_features.iter().cloned());
-
-        match spec {
-            ImportSpec::Module { alias } => {
-                if let Some(namespace) = alias {
-                    self.add_module_namespace(namespace, module_symbols, module_path, span)?;
-                }
-            }
-            ImportSpec::Item { item, alias } => {
-                let symbol_name = alias.as_ref().unwrap_or(item);
-                self.import_single_symbol(&module_symbols, item, symbol_name, module_path, span)?;
-            }
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    let symbol_name = alias.as_ref().unwrap_or(item);
-                    self.import_single_symbol(
-                        &module_symbols,
-                        item,
-                        symbol_name,
-                        module_path,
-                        span,
-                    )?;
-                }
-            }
-            ImportSpec::Wildcard => {
-                self.import_all_symbols(&module_symbols, module_path, span)?;
-            }
-        }
-
-        resolver
-            .borrow_mut()
-            .cache_module(module_path, module_nodes.clone());
-        resolver.borrow_mut().finish_import(module_path);
-
-        self.all_module_asts
-            .insert(module_path.to_string(), module_nodes);
-
-        if !self.module_dependencies.contains(&module_path.to_string()) {
-            self.module_dependencies.push(module_path.to_string());
-        }
-
-        Ok(())
-    }
-
-    fn analyze_statement(
-        &mut self,
-        stmt: &StatementNode,
-        mut files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        match &stmt.kind {
-            StatementKind::AutoDecl(name, _, expr) => {
-                self.analyze_auto_decl_statement(name, expr, stmt.span)?;
-            }
-            StatementKind::TypedDecl(name, type_node, expr) => {
-                self.analyze_typed_decl_statement(name, type_node, expr, stmt.span)?;
-            }
-            StatementKind::Expression(expr) => {
-                self.analyze_expression(expr)?;
-            }
-            StatementKind::Block(stmts) => {
-                self.symbol_table.push_scope()?;
-                self.analyze_block(stmts, files.as_deref_mut())?;
-                self.symbol_table.pop_scope()?;
-            }
-            StatementKind::ConstDecl(name, type_node, expr) => {
-                self.analyze_const_decl_statement(name, type_node, expr, stmt.span)?;
-            }
-            StatementKind::If {
-                cond,
-                then_block,
-                else_block,
-            } => {
-                self.analyze_if_statement(cond, then_block, else_block, files.as_deref_mut())?;
-            }
-            StatementKind::For {
-                var, iter, body, ..
-            } => {
-                self.analyze_for_statement(var, iter, body, stmt.span, files.as_deref_mut())?;
-            }
-            StatementKind::While { cond, body } => {
-                self.analyze_while_statement(cond, body, files.as_deref_mut())?;
-            }
-            StatementKind::Match { expr, arms } => {
-                self.analyze_match_statement(expr, arms, files.as_deref_mut())?;
-            }
-            StatementKind::Return(Some(expr)) => {
-                self.analyze_return_with_value(expr)?;
-            }
-            StatementKind::Return(None) => {
-                self.analyze_return_without_value(stmt.span)?;
-            }
-            StatementKind::Import { module_path, spec } => {
-                self.analyze_import_statement(module_path, spec, stmt.span, files)?;
-            }
-            StatementKind::Function(func) => {
-                self.analyze_function(func, None)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    // Check if a match expression covers all variants of an enum
-    fn check_match_exhaustiveness(
-        &self,
-        expr_type: &Type,
-        arms: &[crate::ast::MatchArm],
-        expr_span: Span,
-    ) -> Result<(), SemanticError> {
-        match expr_type {
-            Type::Result(_, _) => self.check_result_exhaustiveness(arms, expr_span),
-            Type::Named(type_name, _) => {
-                self.check_named_enum_exhaustiveness(type_name, arms, expr_type, expr_span)
-            }
-            Type::Optional(_) => self.check_optional_exhaustiveness(arms, expr_span),
-            _ => self.require_wildcard_pattern(arms, expr_type, expr_span),
-        }
-    }
-
-    fn check_result_exhaustiveness(
-        &self,
-        arms: &[crate::ast::MatchArm],
-        expr_span: Span,
-    ) -> Result<(), SemanticError> {
-        let has_ok = arms.iter().any(|arm| {
-            arm.guard.is_none()
-                && matches!(&arm.pattern, PatternNode::EnumVariant { name, .. } if name == "ok")
-        });
-        let has_err = arms.iter().any(|arm| {
-            arm.guard.is_none()
-                && matches!(&arm.pattern, PatternNode::EnumVariant { name, .. } if name == "err")
-        });
-        let has_wildcard = arms
-            .iter()
-            .any(|arm| matches!(&arm.pattern, PatternNode::Wildcard));
-
-        if has_wildcard || (has_ok && has_err) {
-            return Ok(());
-        }
-
-        let mut missing = Vec::new();
-        if !has_ok {
-            missing.push("ok");
-        }
-        if !has_err {
-            missing.push("err");
-        }
-        Err(SemanticError::with_help(
-            format!(
-                "Non-exhaustive match: missing pattern{} for Result: {}",
-                if missing.len() > 1 { "s" } else { "" },
-                missing.join(", ")
-            ),
-            expr_span,
-            format!(
-                "Add match arm{} for: {}, or add a wildcard '_' pattern to cover all remaining cases",
-                if missing.len() > 1 { "s" } else { "" },
-                missing.join(", ")
-            ),
-        ))
-    }
-
-    fn check_optional_exhaustiveness(
-        &self,
-        arms: &[crate::ast::MatchArm],
-        expr_span: Span,
-    ) -> Result<(), SemanticError> {
-        let has_some = arms.iter().any(|arm| {
-            arm.guard.is_none()
-                && matches!(&arm.pattern, PatternNode::EnumVariant { name, .. } if name == "some")
-        });
-        let has_none = arms.iter().any(|arm| {
-            arm.guard.is_none()
-                && matches!(&arm.pattern, PatternNode::EnumVariant { name, .. } if name == "none")
-        });
-        let has_wildcard = arms
-            .iter()
-            .any(|arm| matches!(&arm.pattern, PatternNode::Wildcard));
-
-        if has_wildcard || (has_some && has_none) {
-            return Ok(());
-        }
-
-        let mut missing = Vec::new();
-        if !has_some {
-            missing.push("some");
-        }
-        if !has_none {
-            missing.push("none");
-        }
-        Err(SemanticError::with_help(
-            format!(
-                "Non-exhaustive match: missing pattern{} for Optional: {}",
-                if missing.len() > 1 { "s" } else { "" },
-                missing.join(", ")
-            ),
-            expr_span,
-            format!(
-                "Add match arm{} for: {}, or add a wildcard '_' pattern",
-                if missing.len() > 1 { "s" } else { "" },
-                missing.join(", ")
-            ),
-        ))
-    }
-
-    fn check_named_enum_exhaustiveness(
-        &self,
-        type_name: &str,
-        arms: &[crate::ast::MatchArm],
-        expr_type: &Type,
-        expr_span: Span,
-    ) -> Result<(), SemanticError> {
-        let Some(symbol) = self.symbol_table.lookup(type_name) else {
-            return self.require_wildcard_pattern(arms, expr_type, expr_span);
-        };
-        let Some(variant_names) = &symbol.variants else {
-            return self.require_wildcard_pattern(arms, expr_type, expr_span);
-        };
-
-        let (covered, has_wildcard) = self.collect_enum_covered_variants(arms, variant_names);
-
-        if has_wildcard {
-            return Ok(());
-        }
-
-        let uncovered: Vec<&String> = variant_names
-            .iter()
-            .filter(|v| !covered.contains(*v))
-            .collect();
-
-        if !uncovered.is_empty() {
-            let uncovered_list = uncovered
-                .iter()
-                .map(|s| s.as_str())
-                .collect::<Vec<_>>()
-                .join(", ");
-            return Err(SemanticError::with_help(
-                format!(
-                    "Non-exhaustive match: missing variant{} of '{}': {}",
-                    if uncovered.len() > 1 { "s" } else { "" },
-                    type_name,
-                    uncovered_list
-                ),
-                expr_span,
-                format!(
-                    "Add match arm{} for: {}, or add a wildcard '_' pattern",
-                    if uncovered.len() > 1 { "s" } else { "" },
-                    uncovered_list
-                ),
-            ));
-        }
-        Ok(())
-    }
-
-    fn collect_enum_covered_variants(
-        &self,
-        arms: &[crate::ast::MatchArm],
-        variant_names: &[String],
-    ) -> (std::collections::HashSet<String>, bool) {
-        let mut covered: std::collections::HashSet<String> = std::collections::HashSet::new();
-        let mut has_wildcard = false;
-
-        for arm in arms {
-            match &arm.pattern {
-                PatternNode::Wildcard => {
-                    has_wildcard = true;
-                    break;
-                }
-                PatternNode::EnumVariant { name, .. } => {
-                    if arm.guard.is_none() {
-                        covered.insert(name.clone());
-                    }
-                }
-                PatternNode::Identifier(name) => {
-                    if arm.guard.is_none() && variant_names.contains(name) {
-                        covered.insert(name.clone());
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        (covered, has_wildcard)
-    }
-
-    fn require_wildcard_pattern(
-        &self,
-        arms: &[crate::ast::MatchArm],
-        expr_type: &Type,
-        expr_span: Span,
-    ) -> Result<(), SemanticError> {
-        let has_wildcard = arms
-            .iter()
-            .any(|arm| matches!(arm.pattern, PatternNode::Wildcard));
-        if !has_wildcard {
-            return Err(SemanticError::with_help(
-                format!("Non-exhaustive match on type '{}'", format_type(expr_type)),
-                expr_span,
-                "Add a wildcard '_' pattern as the last match arm to handle all remaining cases",
-            ));
-        }
-        Ok(())
-    }
-
-    // Set types for pattern variables based on expected type
-    // Uses expr.span for pattern errors to point to the match expression
-    fn set_pattern_types(
-        &mut self,
-        pattern: &PatternNode,
-        expected_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        match pattern {
-            PatternNode::Identifier(name) => {
-                self.handle_identifier_pattern(name, expected_type, span)?;
-            }
-            PatternNode::EnumVariant { name, args } => {
-                self.handle_enum_variant_pattern(name, args, expected_type, span)?;
-            }
-            PatternNode::Literal(lit) => {
-                self.handle_literal_pattern(lit, expected_type, span)?;
-            }
-            PatternNode::List { elements, rest } => {
-                self.handle_list_pattern(elements, rest, expected_type, span)?;
-            }
-            PatternNode::Wildcard => {} // no binding
-        }
-        Ok(())
-    }
-
-    fn handle_identifier_pattern(
-        &mut self,
-        name: &str,
-        expected_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let is_constant = self
-            .symbol_table
-            .lookup(name)
-            .map(|s| s.kind == SymbolKind::Constant)
-            .unwrap_or(false);
-
-        if is_constant {
-            let const_type = self
-                .symbol_table
-                .lookup(name)
-                .and_then(|s| s.type_.clone())
-                .ok_or_else(|| {
-                    SemanticError::new(format!("Constant '{}' has no type information", name), span)
-                })?;
-            self.check_type_compatibility(&const_type, expected_type, span)?;
-        } else {
-            self.symbol_table.add_symbol(
-                name,
-                Symbol {
-                    kind: SymbolKind::Variable,
-                    span,
-                    type_: Some(expected_type.clone()),
-                    interfaces: std::collections::HashMap::new(),
-                    methods: std::collections::HashMap::new(),
-                    fields: std::collections::HashMap::new(),
-                    type_params: Vec::new(),
-                    original_name: None,
-                    llvm_name: None,
-                    default_param_count: 0,
-                    variants: None,
-                },
-            )?;
-        }
-        Ok(())
-    }
-
-    fn handle_enum_variant_pattern(
-        &mut self,
-        name: &str,
-        args: &[PatternNode],
-        expected_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        match expected_type {
-            Type::Optional(inner) => self.match_optional_variant(name, args, inner, span),
-            Type::Result(ok_type, err_type) => {
-                self.match_result_variant(name, args, ok_type, err_type, span)
-            }
-            Type::Named(enum_name, _) => {
-                let symbol = self
-                    .symbol_table
-                    .lookup(enum_name)
-                    .ok_or_else(|| self.undefined_symbol_error("type", enum_name, span))?;
-                self.match_enum_variant(name, args, enum_name, &symbol, span)
-            }
-            _ => Err(SemanticError::with_help(
-                format!(
-                    "Enum variant patterns are not supported for type {}",
-                    format_type(expected_type)
-                ),
-                span,
-                "Variant patterns can only be used with enum, Optional, or Result types",
-            )),
-        }
-    }
-
-    fn match_optional_variant(
-        &mut self,
-        name: &str,
-        args: &[PatternNode],
-        inner: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        if name == "some" && args.len() == 1 {
-            self.set_pattern_types(&args[0], inner, span)?;
-            Ok(())
-        } else if name == "none" && args.is_empty() {
-            Ok(())
-        } else {
-            Err(SemanticError::with_help(
-                format!(
-                    "Pattern '{}' does not match type {}",
-                    name,
-                    format_type(&Type::Optional(Box::new(inner.clone())))
-                ),
-                span,
-                "Optional values can only be matched with Some(value) or None",
-            ))
-        }
-    }
-
-    fn match_result_variant(
-        &mut self,
-        name: &str,
-        args: &[PatternNode],
-        ok_type: &Type,
-        err_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        if name == "ok" && args.len() == 1 {
-            self.set_pattern_types(&args[0], ok_type, span)?;
-            Ok(())
-        } else if name == "err" && args.len() == 1 {
-            self.set_pattern_types(&args[0], err_type, span)?;
-            Ok(())
-        } else {
-            Err(SemanticError::with_help(
-                format!(
-                    "Pattern '{}' does not match type {}",
-                    name,
-                    format_type(&Type::Result(
-                        Box::new(ok_type.clone()),
-                        Box::new(err_type.clone())
-                    ))
-                ),
-                span,
-                "Result values can only be matched with Ok(value) or Err(value)",
-            ))
-        }
-    }
-
-    fn handle_literal_pattern(
-        &mut self,
-        lit: &crate::ast::LiteralNode,
-        expected_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let literal_type = match lit {
-            crate::ast::LiteralNode::Integer(_) => Type::Primitive(PrimitiveType::Int),
-            crate::ast::LiteralNode::Float(_) => Type::Primitive(PrimitiveType::Float),
-            crate::ast::LiteralNode::String(_) => Type::Primitive(PrimitiveType::Str),
-            crate::ast::LiteralNode::Boolean(_) => Type::Primitive(PrimitiveType::Bool),
-            crate::ast::LiteralNode::Char(_) => Type::Primitive(PrimitiveType::Char),
-        };
-        self.check_type_compatibility(&literal_type, expected_type, span)
-    }
-
-    fn handle_list_pattern(
-        &mut self,
-        elements: &[PatternNode],
-        rest: &Option<Box<PatternNode>>,
-        expected_type: &Type,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let inner_type = match expected_type {
-            Type::List(inner) => (**inner).clone(),
-            Type::EmptyList => Type::Void,
-            _ => {
-                return Err(SemanticError::with_help(
-                    format!(
-                        "List pattern cannot match type {}",
-                        format_type(expected_type)
-                    ),
-                    span,
-                    "List patterns (e.g. [head, ...rest]) can only match list types",
-                ));
-            }
-        };
-        for elem in elements {
-            self.set_pattern_types(elem, &inner_type, span)?;
-        }
-        if let Some(rest_pat) = rest {
-            let rest_type = Type::List(Box::new(inner_type));
-            self.set_pattern_types(rest_pat, &rest_type, span)?;
-        }
-        Ok(())
-    }
-
-    fn match_enum_variant(
-        &mut self,
-        name: &str,
-        args: &[PatternNode],
-        enum_name: &str,
-        symbol: &crate::semantics::Symbol,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let sig = symbol.methods.get(name).ok_or_else(|| {
-            let available_variants: Vec<&String> = symbol.methods.keys().collect();
-            if available_variants.is_empty() {
-                SemanticError::new(
-                    format!("Unknown variant '{}' for enum '{}'", name, enum_name),
-                    span,
-                )
-            } else {
-                SemanticError::with_help(
-                    format!("Unknown variant '{}' for enum '{}'", name, enum_name),
-                    span,
-                    format!(
-                        "Available variants: {}",
-                        available_variants
-                            .iter()
-                            .map(|s| s.as_str())
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                )
-            }
-        })?;
-        if args.len() != sig.params.len() {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Variant '{}' expects {} argument{}, but pattern provides {}",
-                    name,
-                    sig.params.len(),
-                    if sig.params.len() == 1 { "" } else { "s" },
-                    args.len()
-                ),
-                span,
-                format!(
-                    "Match the variant with exactly {} argument{}",
-                    sig.params.len(),
-                    if sig.params.len() == 1 { "" } else { "s" }
-                ),
-            ));
-        }
-        for (arg, param_type) in args.iter().zip(&sig.params) {
-            self.set_pattern_types(arg, param_type, span)?;
-        }
-        Ok(())
-    }
-
-    // analyze a pattern (for match arms).
-    #[allow(clippy::only_used_in_recursion)]
-    fn analyze_pattern(&mut self, pattern: &PatternNode) -> Result<(), SemanticError> {
-        match pattern {
-            PatternNode::Identifier(_) => {} // already added in set_pattern_types
-            PatternNode::EnumVariant { args, .. } => {
-                for arg in args {
-                    self.analyze_pattern(arg)?;
-                }
-            }
-            PatternNode::List { elements, rest } => {
-                for elem in elements {
-                    self.analyze_pattern(elem)?;
-                }
-                if let Some(rest_pat) = rest {
-                    self.analyze_pattern(rest_pat)?;
-                }
-            }
-            PatternNode::Literal(_) => {} // literals don't bind variables
-            PatternNode::Wildcard => {}   // no binding
-        }
-        Ok(())
-    }
-
-    // analyze an expression.
-    fn analyze_expression(&mut self, expr: &ExpressionNode) -> Result<(), SemanticError> {
-        match &expr.kind {
-            ExpressionKind::Identifier(name) => self.analyze_identifier_expr(name, expr),
-            ExpressionKind::Literal(_) => Ok(()),
-            ExpressionKind::None => Ok(()),
-            ExpressionKind::Binary { left, right, .. } => {
-                self.analyze_expression(left)?;
-                self.analyze_expression(right)?;
-                let _ = self.get_expression_type(expr)?;
-                Ok(())
-            }
-            ExpressionKind::Unary {
-                expr,
-                op,
-                op_span,
-                postfix: _,
-            } => self.analyze_unary_expr(expr, op, *op_span),
-            ExpressionKind::Call { func, args } => self.analyze_call_expr(expr, func, args),
-            ExpressionKind::FieldAccess { expr, .. } => {
-                self.analyze_expression(expr)?;
-                let _ = self.get_expression_type(expr)?;
-                Ok(())
-            }
-            ExpressionKind::ListAccess { expr, index } => {
-                self.analyze_list_access_expr(expr, index)
-            }
-            ExpressionKind::ListLiteral(elements) => self.analyze_list_literal_expr(elements),
-            ExpressionKind::MapLiteral { entries, .. } => {
-                self.analyze_map_literal_expr(expr, entries)
-            }
-            ExpressionKind::SetOrMapLiteral(elements) => self.analyze_set_literal_expr(elements),
-            ExpressionKind::TupleLiteral(elements) => {
-                self.analyze_tuple_literal_expr(expr, elements)
-            }
-            ExpressionKind::If {
-                cond,
-                then_expr,
-                else_expr,
-            } => self.analyze_if_expr(cond, then_expr, else_expr),
-            ExpressionKind::Lambda {
-                params,
-                return_type,
-                body,
-            } => self.analyze_lambda_expr(expr, params, return_type, body),
-            ExpressionKind::GenericType(name, type_args) => {
-                self.analyze_generic_type_expr(expr, name, type_args)
-            }
-        }
-    }
-
-    fn analyze_identifier_expr(
-        &mut self,
-        name: &str,
-        expr: &ExpressionNode,
-    ) -> Result<(), SemanticError> {
-        if name == "self" {
-            return self.analyze_self_identifier(expr);
-        }
-        let mut exists_like =
-            self.symbol_table.exists(name) || self.get_builtin_sig(name).is_some();
-        if !exists_like {
-            exists_like = self.check_stdlib_imports_for_class(name);
-        }
-        if !exists_like {
-            // Only suggest the collection constructor when no similar symbol
-            // exists. If a similar symbol does exist, the standard "did you
-            // mean?" diagnostic from undefined_symbol_error is more helpful
-            // for catching typos like a variable named `listing` vs `list`.
-            if self.symbol_table.find_similar(name).is_none()
-                && let Some(help) = collection_new_hint(name)
-            {
-                return Err(SemanticError::with_help(
-                    format!("Undefined variable '{}'", name),
-                    expr.span,
-                    help,
-                ));
-            }
-            return Err(self.undefined_symbol_error("variable", name, expr.span));
-        }
-        Ok(())
-    }
-
-    fn analyze_self_identifier(&self, expr: &ExpressionNode) -> Result<(), SemanticError> {
-        if self.is_in_static_method {
-            return Err(SemanticError::with_help(
-                "Cannot use 'self' in a common method",
-                expr.span,
-                "Common (static) methods do not have access to 'self'. Remove the 'common' modifier or access the class through a parameter instead.",
-            ));
-        }
-        if self.current_self_type.is_none() {
-            return Err(SemanticError::with_help(
-                "Cannot use 'self' outside of a method",
-                expr.span,
-                "'self' is only available inside instance methods of a class",
-            ));
-        }
-        Ok(())
-    }
-
-    fn check_stdlib_imports_for_class(&self, name: &str) -> bool {
-        let stdlib_names: std::collections::HashSet<String> = std_module_registry()
-            .keys()
-            .filter_map(|s| s.strip_prefix("std.").map(|name| name.to_string()))
-            .collect();
-        for (ns, module_symbols) in &self.imported_symbols {
-            if !stdlib_names.contains(ns) {
-                continue;
-            }
-            if let Some(sym) = module_symbols.get(name)
-                && matches!(sym.kind, SymbolKind::Class)
-            {
-                return true;
-            }
-        }
-        false
-    }
-
-    fn analyze_unary_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        op: &UnaryOp,
-        op_span: Span,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(expr)?;
-        let operand_type = self.get_expression_type(expr)?;
-        match op {
-            UnaryOp::Not => self.check_not_operator_type(&operand_type, op_span),
-            UnaryOp::Neg => self.check_neg_operator_type(&operand_type, op_span),
-            UnaryOp::Ref => Ok(()),
-            UnaryOp::Incr | UnaryOp::Decr => {
-                self.check_incr_decr_operator_type(&operand_type, op_span)?;
-                self.check_incr_decr_const_modification(expr, op_span)
-            }
-            _ => Ok(()),
-        }
-    }
-
-    fn check_not_operator_type(
-        &self,
-        operand_type: &Type,
-        op_span: Span,
-    ) -> Result<(), SemanticError> {
-        if !matches!(
-            operand_type,
-            Type::Primitive(crate::ast::PrimitiveType::Bool)
-        ) {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Logical 'not' operator requires a boolean operand, found {}",
-                    format_type(operand_type)
-                ),
-                op_span,
-                "The '!' operator can only be applied to bool values",
-            ));
-        }
-        Ok(())
-    }
-
-    fn check_neg_operator_type(
-        &self,
-        operand_type: &Type,
-        op_span: Span,
-    ) -> Result<(), SemanticError> {
-        if !matches!(
-            operand_type,
-            Type::Primitive(crate::ast::PrimitiveType::Int)
-                | Type::Primitive(crate::ast::PrimitiveType::Float)
-        ) {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Negation operator '-' requires a numeric operand, found {}",
-                    format_type(operand_type)
-                ),
-                op_span,
-                "The unary '-' operator can only be applied to int or float values",
-            ));
-        }
-        Ok(())
-    }
-
-    fn check_incr_decr_operator_type(
-        &self,
-        operand_type: &Type,
-        op_span: Span,
-    ) -> Result<(), SemanticError> {
-        if !matches!(
-            operand_type,
-            Type::Primitive(crate::ast::PrimitiveType::Int)
-        ) {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Increment/decrement operators require an int operand, found {}",
-                    format_type(operand_type)
-                ),
-                op_span,
-                "The '++' and '--' operators can only be applied to int variables",
-            ));
-        }
-        Ok(())
-    }
-
-    fn check_incr_decr_const_modification(
-        &mut self,
-        expr: &ExpressionNode,
-        op_span: Span,
-    ) -> Result<(), SemanticError> {
-        if let crate::ast::ExpressionKind::Identifier(name) = &expr.kind
-            && let Some(symbol) = self.symbol_table.lookup(name)
-            && symbol.kind == SymbolKind::Constant
-        {
-            return Err(SemanticError::with_help(
-                format!("Cannot modify constant '{}'", name),
-                op_span,
-                "Constants cannot be modified after initialization",
-            ));
-        }
-
-        if let crate::ast::ExpressionKind::FieldAccess {
-            expr: obj_expr,
-            field,
-        } = &expr.kind
-        {
-            let obj_type = self.get_expression_type(obj_expr)?;
-            if let Type::Named(class_name, _) = &obj_type
-                && let Some(symbol) = self.symbol_table.lookup(class_name)
-                && let Some((_field_type, is_const)) = symbol.fields.get(field)
-                && *is_const
-            {
-                return Err(SemanticError::with_help(
-                    format!("Cannot modify const field '{}'", field),
-                    op_span,
-                    "Const fields cannot be modified after initialization. Remove the 'const' modifier from the field declaration if mutation is needed.",
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn analyze_call_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        func: &ExpressionNode,
-        args: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        if let ExpressionKind::Identifier(name) = &func.kind
-            && !self.symbol_table.exists(name)
-            && self.get_builtin_sig(name).is_none()
-        {
-            return Err(self.undefined_symbol_error("function", name, func.span));
-        }
-
-        self.analyze_expression(func)?;
-        for arg in args {
-            self.analyze_expression(arg)?;
-        }
-
-        if let ExpressionKind::Identifier(name) = &func.kind
-            && name == "some"
-        {
-            self.check_some_call_args(expr, args)?;
-        }
-
-        let _ = self.get_expression_type(expr)?;
-        Ok(())
-    }
-
-    fn check_some_call_args(
-        &mut self,
-        expr: &ExpressionNode,
-        args: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        if args.len() != 1 {
-            return Err(SemanticError::with_help(
-                format!("Some() takes exactly 1 argument, got {}", args.len()),
-                expr.span,
-                "Wrap a single value in Some(), e.g. Some(42)",
-            ));
-        }
-        let arg_type = self.get_expression_type(&args[0])?;
-        if let Type::Optional(_) = arg_type {
-            return Err(SemanticError::with_help(
-                "Some() cannot wrap an Optional value",
-                expr.span,
-                "The argument to Some() must not be Optional. Remove the nested Some() or unwrap the inner value first.",
-            ));
-        }
-        Ok(())
-    }
-
-    fn analyze_list_access_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        index: &ExpressionNode,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(expr)?;
-        self.analyze_expression(index)?;
-        let target_type = self.get_expression_type(expr)?;
-        let index_type = self.get_expression_type(index)?;
-        match &target_type {
-            Type::List(_) => {
-                if !matches!(index_type, Type::Primitive(crate::ast::PrimitiveType::Int)) {
-                    return Err(SemanticError::with_help(
-                        format!(
-                            "List index must be an integer, found {}",
-                            format_type(&index_type)
-                        ),
-                        index.span,
-                        "Lists can only be indexed with integer values, e.g. myList[0]",
-                    ));
-                }
-            }
-            Type::Map(expected_key_type, _) => {
-                if index_type != **expected_key_type {
-                    return Err(SemanticError::with_help(
-                        format!(
-                            "Map key type mismatch: expected {}, found {}",
-                            format_type(expected_key_type),
-                            format_type(&index_type)
-                        ),
-                        index.span,
-                        format!(
-                            "This map has keys of type {}",
-                            format_type(expected_key_type)
-                        ),
-                    ));
-                }
-            }
-            Type::EmptyMap => {
-                return Err(SemanticError::with_help(
-                    "Cannot index empty map",
-                    expr.span,
-                    "The map type is unknown. Provide type annotations or add entries to the map literal.",
-                ));
-            }
-            _ => {
-                return Err(SemanticError::with_help(
-                    "Cannot index non-list type",
-                    expr.span,
-                    "Only lists and maps can be indexed with '[]'. Examples: my_list[0], my_map['key']",
-                ));
-            }
-        }
-        Ok(())
-    }
-
-    fn analyze_list_literal_expr(
-        &mut self,
-        elements: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        for elem in elements {
-            self.analyze_expression(elem)?;
-        }
-        if !elements.is_empty() {
-            let first_type = self.get_expression_type(&elements[0])?;
-            for elem in &elements[1..] {
-                let elem_type = self.get_expression_type(elem)?;
-                if elem_type != first_type {
-                    return Err(SemanticError::with_help(
-                        format!(
-                            "List element type mismatch: expected {}, found {}",
-                            format_type(&first_type),
-                            format_type(&elem_type)
-                        ),
-                        elem.span,
-                        "All elements in a list literal must have the same type",
-                    ));
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn analyze_map_literal_expr(
-        &mut self,
-        _expr: &ExpressionNode,
-        entries: &[(ExpressionNode, ExpressionNode)],
-    ) -> Result<(), SemanticError> {
-        for (key, value) in entries {
-            self.analyze_expression(key)?;
-            self.analyze_expression(value)?;
-        }
-        if !entries.is_empty() {
-            let (first_key, first_value) = &entries[0];
-            let key_type = self.get_expression_type(first_key)?;
-            self.check_map_key_hashable(first_key, &key_type)?;
-
-            let value_type = self.get_expression_type(first_value)?;
-            for (key, value) in &entries[1..] {
-                self.check_map_entry_type_consistency(key, value, &key_type, &value_type)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn check_map_key_hashable(
-        &self,
-        key_expr: &ExpressionNode,
-        key_type: &Type,
-    ) -> Result<(), SemanticError> {
-        let is_hashable = matches!(key_type, Type::Primitive(_));
-        if !is_hashable {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Map keys must be a hashable type, found '{}'",
-                    format_type(key_type)
-                ),
-                key_expr.span,
-                "Only primitive types (int, float, string, bool, char) can be used as map keys",
-            ));
-        }
-        Ok(())
-    }
-
-    fn check_map_entry_type_consistency(
-        &mut self,
-        key: &ExpressionNode,
-        value: &ExpressionNode,
-        expected_key: &Type,
-        expected_value: &Type,
-    ) -> Result<(), SemanticError> {
-        let k_type = self.get_expression_type(key)?;
-        self.check_map_key_hashable(key, &k_type)?;
-
-        let v_type = self.get_expression_type(value)?;
-        if k_type != *expected_key {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Map key type mismatch: expected {}, found {}",
-                    format_type(expected_key),
-                    format_type(&k_type)
-                ),
-                key.span,
-                "All keys in a map literal must have the same type",
-            ));
-        }
-        if v_type != *expected_value {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Map value type mismatch: expected {}, found {}",
-                    format_type(expected_value),
-                    format_type(&v_type)
-                ),
-                value.span,
-                "All values in a map literal must have the same type",
-            ));
-        }
-        Ok(())
-    }
-
-    fn analyze_set_literal_expr(
-        &mut self,
-        elements: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        for elem in elements {
-            self.analyze_expression(elem)?;
-        }
-        if !elements.is_empty() {
-            let first_type = self.get_expression_type(&elements[0])?;
-            for elem in &elements[1..] {
-                let elem_type = self.get_expression_type(elem)?;
-                if elem_type != first_type {
-                    return Err(SemanticError::with_help(
-                        format!(
-                            "Set element type mismatch: expected {}, found {}",
-                            format_type(&first_type),
-                            format_type(&elem_type)
-                        ),
-                        elem.span,
-                        "All elements in a set literal must have the same type",
-                    ));
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn analyze_tuple_literal_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        elements: &[ExpressionNode],
-    ) -> Result<(), SemanticError> {
-        if elements.len() != 2 {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Tuple must have exactly 2 elements, found {}",
-                    elements.len()
-                ),
-                expr.span,
-                "Tuples in Mux are pairs with exactly two elements, e.g. (1, 2)",
-            ));
-        }
-        for elem in elements {
-            self.analyze_expression(elem)?;
-        }
-        Ok(())
-    }
-
-    fn analyze_if_expr(
-        &mut self,
-        cond: &ExpressionNode,
-        then_expr: &ExpressionNode,
-        else_expr: &ExpressionNode,
-    ) -> Result<(), SemanticError> {
-        self.analyze_expression(cond)?;
-        self.analyze_expression(then_expr)?;
-        self.analyze_expression(else_expr)?;
-        let cond_type = self.get_expression_type(cond)?;
-        if !matches!(cond_type, Type::Primitive(crate::ast::PrimitiveType::Bool)) {
-            return Err(SemanticError::with_help(
-                format!(
-                    "If condition must be boolean, found {}",
-                    format_type(&cond_type)
-                ),
-                cond.span,
-                "The condition in an if expression must evaluate to a bool value",
-            ));
-        }
-        Ok(())
-    }
-
-    fn analyze_lambda_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        params: &[Param],
-        return_type: &TypeNode,
-        body: &[StatementNode],
-    ) -> Result<(), SemanticError> {
-        let mut local_vars = std::collections::HashSet::new();
-        for param in params {
-            local_vars.insert(param.name.clone());
-        }
-
-        self.symbol_table.push_scope()?;
-
-        let lambda_return_type = self.resolve_type(return_type)?;
-        let prev_return_type = self.current_return_type.clone();
-        self.current_return_type = Some(lambda_return_type.clone());
-
-        for param in params {
-            let param_type = self.resolve_type(&param.type_)?;
-            self.symbol_table.add_symbol(
-                &param.name,
-                Symbol {
-                    kind: SymbolKind::Variable,
-                    span: param.type_.span,
-                    type_: Some(param_type),
-                    interfaces: std::collections::HashMap::new(),
-                    methods: std::collections::HashMap::new(),
-                    fields: std::collections::HashMap::new(),
-                    type_params: Vec::new(),
-                    original_name: None,
-                    llvm_name: None,
-                    default_param_count: 0,
-                    variants: None,
-                },
-            )?;
-        }
-
-        self.analyze_block(body, None)?;
-
-        self.check_lambda_return_paths(expr, body, &lambda_return_type)?;
-
-        self.current_return_type = prev_return_type;
-        let captures = self.find_free_variables_in_block(body, &local_vars)?;
-        self.lambda_captures.insert(expr.span, captures);
-
-        self.symbol_table.pop_scope()?;
-        Ok(())
-    }
-
-    fn check_lambda_return_paths(
-        &mut self,
-        expr: &ExpressionNode,
-        body: &[StatementNode],
-        lambda_return_type: &Type,
-    ) -> Result<(), SemanticError> {
-        if body.is_empty() || !self.all_paths_return(body) {
-            let (msg, help) = if matches!(lambda_return_type, Type::Void) {
-                (
-                    "Lambda must end with an explicit 'return' statement on all code paths"
-                        .to_string(),
-                    "Add a 'return' statement at the end of every code path in the lambda body"
-                        .to_string(),
-                )
-            } else {
-                (
-                    format!(
-                        "Lambda must return a value of type '{}' on all code paths",
-                        format_type(lambda_return_type)
-                    ),
-                    "Add a return statement at the end of every branch in the lambda body"
-                        .to_string(),
-                )
-            };
-            return Err(SemanticError::with_help(msg, expr.span, help));
-        }
-        if let Some(last_stmt) = body.last()
-            && let StatementKind::Return(Some(ret_expr)) = &last_stmt.kind
-        {
-            let actual_type = self.get_expression_type(ret_expr)?;
-            self.check_type_compatibility(lambda_return_type, &actual_type, ret_expr.span)?;
-        }
-        Ok(())
-    }
-
-    fn analyze_generic_type_expr(
-        &mut self,
-        expr: &ExpressionNode,
-        name: &str,
-        type_args: &[TypeNode],
-    ) -> Result<(), SemanticError> {
-        if name == "tuple" {
-            return self.check_tuple_type_args(expr, type_args);
-        }
-        // Only suggest the collection constructor when no similar symbol
-        // exists. A typo like `lisp<int>.new()` should surface the
-        // "did you mean 'list'?" suggestion instead.
-        if self.symbol_table.find_similar(name).is_none()
-            && let Some(help) = collection_new_hint(name)
-        {
-            return Err(SemanticError::with_help(
-                format!("Undefined type '{}'", name),
-                expr.span,
-                help,
-            ));
-        }
-        if let Some((module_name, type_name)) = name.split_once('.') {
-            let module_symbols = self
-                .imported_symbols
-                .get(module_name)
-                .ok_or_else(|| self.undefined_symbol_error("module", module_name, expr.span))?;
-            let _ = module_symbols
-                .get(type_name)
-                .ok_or_else(|| self.undefined_symbol_error("type", type_name, expr.span))?;
-        } else if !self.symbol_table.exists(name) {
-            return Err(self.undefined_symbol_error("type", name, expr.span));
-        }
-        Ok(())
-    }
-
-    fn check_tuple_type_args(
-        &self,
-        expr: &ExpressionNode,
-        type_args: &[TypeNode],
-    ) -> Result<(), SemanticError> {
-        if type_args.len() != 2 {
-            return Err(SemanticError::with_help(
-                format!(
-                    "Tuple type requires exactly 2 type arguments, got {}",
-                    type_args.len()
-                ),
-                expr.span,
-                "Tuples in Mux are pairs, e.g. tuple<int, string>",
-            ));
-        }
-        for arg in type_args {
-            self.resolve_type(arg)?;
-        }
-        Ok(())
-    }
-
-    fn infer_literal_type(&self, expr: &ExpressionNode) -> Result<Type, SemanticError> {
-        match &expr.kind {
-            ExpressionKind::Literal(lit) => match lit {
-                LiteralNode::Integer(_) => Ok(Type::Primitive(PrimitiveType::Int)),
-                LiteralNode::Float(_) => Ok(Type::Primitive(PrimitiveType::Float)),
-                LiteralNode::String(_) => Ok(Type::Primitive(PrimitiveType::Str)),
-                LiteralNode::Boolean(_) => Ok(Type::Primitive(PrimitiveType::Bool)),
-                LiteralNode::Char(_) => Ok(Type::Primitive(PrimitiveType::Char)),
-            },
-            _ => Err(SemanticError::with_help(
-                "Expected a literal expression",
-                expr.span,
-                "Only literal values (integers, floats, strings, booleans, chars) are allowed here",
-            )),
-        }
-    }
-
-    fn types_compatible(&self, type1: &Type, type2: &Type) -> bool {
-        // Handle type variables - same name means compatible
-        match (type1, type2) {
-            (Type::Variable(v1), Type::Variable(v2)) => v1 == v2,
-            (Type::Generic(g1), Type::Generic(g2)) => g1 == g2,
-            _ => type1 == type2,
-        }
-    }
-
-    // Add module as namespace (import logger as log)
-    fn add_module_namespace(
-        &mut self,
-        namespace: &str,
-        symbols: std::collections::HashMap<String, Symbol>,
-        module_path: &str,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        // Mangle function names in the symbols before storing
-        let module_name_for_mangling = module_path.replace(['.', '/'], "_");
-        let mut mangled_symbols = std::collections::HashMap::new();
-
-        for (name, symbol) in symbols {
-            let mut mangled_symbol = symbol.clone();
-            // Set llvm_name for functions
-            if matches!(symbol.kind, SymbolKind::Function) {
-                mangled_symbol.llvm_name = Some(format!("{}!{}", module_name_for_mangling, name));
-            }
-            mangled_symbols.insert(name, mangled_symbol);
-        }
-
-        // Store module symbols for namespaced access
-        self.imported_symbols
-            .insert(namespace.to_string(), mangled_symbols);
-
-        // Add module symbol to symbol table
-        self.symbol_table.add_symbol(
-            namespace,
-            Symbol {
-                kind: SymbolKind::Import,
-                span,
-                type_: Some(Type::Module(namespace.to_string())),
-                interfaces: std::collections::HashMap::new(),
-                methods: std::collections::HashMap::new(),
-                fields: std::collections::HashMap::new(),
-                type_params: Vec::new(),
-                original_name: None,
-                llvm_name: None,
-                default_param_count: 0,
-                variants: None,
-            },
-        )?;
-
-        Ok(())
-    }
-
-    fn filter_module_export_symbols(
-        &self,
-        symbols: &HashMap<String, Symbol>,
-    ) -> HashMap<String, Symbol> {
-        symbols
-            .iter()
-            .filter(|(_, symbol)| {
-                matches!(
-                    symbol.kind,
-                    SymbolKind::Function
-                        | SymbolKind::Class
-                        | SymbolKind::Interface
-                        | SymbolKind::Enum
-                        | SymbolKind::Constant
-                )
-            })
-            .map(|(name, symbol)| (name.clone(), symbol.clone()))
-            .collect()
-    }
-
-    fn collect_declared_module_symbols(
-        &self,
-        module_nodes: &[AstNode],
-        module_analyzer: &SemanticAnalyzer,
-    ) -> HashMap<String, Symbol> {
-        let global_symbols = module_analyzer.symbol_table.global_scope_symbols();
-        let declared: HashMap<String, Symbol> = module_nodes
-            .iter()
-            .filter_map(Self::declared_symbol_name)
-            .filter_map(|symbol_name| {
-                global_symbols
-                    .get(symbol_name)
-                    .cloned()
-                    .map(|symbol| (symbol_name.to_string(), symbol))
-            })
-            .collect();
-
-        self.filter_module_export_symbols(&declared)
-    }
-
-    fn declared_symbol_name(node: &AstNode) -> Option<&str> {
-        match node {
-            AstNode::Function(func) => Some(func.name.as_str()),
-            AstNode::Class { name, .. }
-            | AstNode::Interface { name, .. }
-            | AstNode::Enum { name, .. } => Some(name.as_str()),
-            AstNode::Statement(stmt) => Self::declared_statement_symbol_name(stmt),
-        }
-    }
-
-    fn declared_statement_symbol_name(stmt: &StatementNode) -> Option<&str> {
-        match &stmt.kind {
-            StatementKind::AutoDecl(name, _, _)
-            | StatementKind::TypedDecl(name, _, _)
-            | StatementKind::ConstDecl(name, _, _) => Some(name.as_str()),
-            StatementKind::Function(func) => Some(func.name.as_str()),
-            _ => None,
-        }
-    }
-
-    fn analyze_imported_module(
-        &mut self,
-        module_nodes: &[AstNode],
-        module_path: &str,
-        resolver: &Rc<RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut Files,
-        span: Span,
-    ) -> Result<HashMap<String, Symbol>, SemanticError> {
-        let mut module_analyzer = SemanticAnalyzer::new_for_module(resolver.clone());
-        module_analyzer.set_current_file(std::path::PathBuf::from(
-            module_path.replace('.', "/") + ".mux",
-        ));
-        let errors = module_analyzer.analyze(module_nodes, Some(files));
-        if !errors.is_empty() {
-            resolver.borrow_mut().finish_import(module_path);
-            let error_messages: Vec<String> = errors.iter().map(|e| e.message.clone()).collect();
-            return Err(SemanticError::with_help(
-                format!("Errors in imported module '{}'", module_path),
-                span,
-                format!(
-                    "Fix the following errors in '{}':\n  {}",
-                    module_path,
-                    error_messages.join("\n  ")
-                ),
-            ));
-        }
-        Ok(self.collect_declared_module_symbols(module_nodes, &module_analyzer))
-    }
-
-    fn mangle_and_import_module_symbols(
-        &mut self,
-        module_symbols: &HashMap<String, Symbol>,
-        module_path: &str,
-    ) -> Result<(), SemanticError> {
-        let module_name_for_mangling = Self::sanitize_module_path(module_path);
-        for (name, symbol) in module_symbols {
-            let name_str = name.as_str();
-            let is_unmangled_builtin_function = matches!(symbol.kind, SymbolKind::Function)
-                && (name_str.starts_with("print")
-                    || name_str.starts_with("read_line")
-                    || name_str.starts_with("range")
-                    || name_str.starts_with("some")
-                    || name_str.starts_with("none")
-                    || name_str.starts_with("ok")
-                    || name_str.starts_with("err"));
-
-            if !is_unmangled_builtin_function && !self.symbol_table.all_symbols.contains_key(name) {
-                let mut mangled_symbol = symbol.clone();
-                if matches!(symbol.kind, SymbolKind::Function) {
-                    mangled_symbol.llvm_name =
-                        Some(format!("{}!{}", module_name_for_mangling, name));
-                }
-                self.symbol_table
-                    .all_symbols
-                    .insert(name.clone(), mangled_symbol);
-            }
-        }
-
-        Ok(())
-    }
-
-    fn handle_import_spec(
-        &mut self,
-        spec: &crate::ast::ImportSpec,
-        module_symbols: &HashMap<String, Symbol>,
-        module_path: &str,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        match spec {
-            ImportSpec::Module { alias } => {
-                if let Some(namespace) = alias {
-                    self.add_module_namespace(
-                        namespace,
-                        module_symbols.clone(),
-                        module_path,
-                        span,
-                    )?;
-                } else if module_path.contains('.') {
-                    let name = module_path
-                        .split('.')
-                        .next_back()
-                        .expect("module path should include at least one segment");
-                    self.add_module_namespace(name, module_symbols.clone(), module_path, span)?;
-                }
-            }
-            ImportSpec::Item { item, alias } => {
-                let symbol_name = alias.as_ref().unwrap_or(item);
-                self.import_single_symbol(module_symbols, item, symbol_name, module_path, span)?;
-            }
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    let symbol_name = alias.as_ref().unwrap_or(item);
-                    self.import_single_symbol(
-                        module_symbols,
-                        item,
-                        symbol_name,
-                        module_path,
-                        span,
-                    )?;
-                }
-            }
-            ImportSpec::Wildcard => {
-                self.import_all_symbols(module_symbols, module_path, span)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn import_module_from_resolver(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        files: &mut Files,
-    ) -> Result<(), SemanticError> {
-        let resolver = self
-            .module_resolver
-            .clone()
-            .ok_or_else(|| SemanticError::new("Module resolver not available", span))?;
-
-        let (has_file, has_directory) = resolver.borrow().check_module_path(module_path);
-        if has_file && has_directory {
-            return Err(SemanticError::with_help(
-                format!("Ambiguous import: '{}'", module_path),
-                span,
-                format!(
-                    "Both {}.mux and {}/ directory exist. Please remove one.",
-                    module_path.replace('.', "/"),
-                    module_path.replace('.', "/")
-                ),
-            ));
-        }
-
-        if has_directory {
-            self.handle_directory_import(module_path, spec, span, resolver, files)?;
-            return Ok(());
-        }
-
-        let module_nodes = resolver
-            .borrow_mut()
-            .resolve_import_path(module_path, self.current_file.as_deref(), files)
-            .map_err(|e| {
-                SemanticError::with_help(
-                    format!("Failed to import module '{}'", module_path),
-                    span,
-                    e.to_string(),
-                )
-            })?;
-
-        let module_symbols =
-            self.analyze_imported_module(&module_nodes, module_path, &resolver, files, span)?;
-        self.mangle_and_import_module_symbols(&module_symbols, module_path)?;
-
-        self.handle_import_spec(spec, &module_symbols, module_path, span)?;
-
-        resolver
-            .borrow_mut()
-            .cache_module(module_path, module_nodes.clone());
-        resolver.borrow_mut().finish_import(module_path);
-        self.all_module_asts
-            .insert(module_path.to_string(), module_nodes);
-
-        if !self.module_dependencies.contains(&module_path.to_string()) {
-            self.module_dependencies.push(module_path.to_string());
-        }
-
-        Ok(())
-    }
-
-    // Handle directory-based module import (e.g., import utils where utils/ is a directory)
-    fn resolve_and_register_submodule(
-        &mut self,
-        submodule_path: &str,
-        namespace: &str,
-        span: Span,
-        resolver: &Rc<RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut Files,
-    ) -> Result<Symbol, SemanticError> {
-        let submodule_nodes = resolver
-            .borrow_mut()
-            .resolve_import_path(submodule_path, self.current_file.as_deref(), files)
-            .map_err(|e| {
-                SemanticError::with_help(
-                    format!("Failed to import submodule '{}'", submodule_path),
-                    span,
-                    e.to_string(),
-                )
-            })?;
-
-        let mut submodule_analyzer = SemanticAnalyzer::new_for_module(resolver.clone());
-        submodule_analyzer.set_current_file(std::path::PathBuf::from(
-            submodule_path.replace('.', "/") + ".mux",
-        ));
-        let errors = submodule_analyzer.analyze(&submodule_nodes, Some(files));
-        if !errors.is_empty() {
-            resolver.borrow_mut().finish_import(submodule_path);
-            let error_messages: Vec<String> = errors.iter().map(|e| e.message.clone()).collect();
-            return Err(SemanticError::with_help(
-                format!("Errors in submodule '{}'", submodule_path),
-                span,
-                format!(
-                    "Fix the following errors in '{}':\n  {}",
-                    submodule_path,
-                    error_messages.join("\n  ")
-                ),
-            ));
-        }
-
-        let submodule_symbols =
-            self.collect_declared_module_symbols(&submodule_nodes, &submodule_analyzer);
-        self.required_runtime_features
-            .extend(submodule_analyzer.required_runtime_features.iter().cloned());
-
-        self.mangle_and_import_module_symbols(&submodule_symbols, submodule_path)?;
-
-        let mangled_submodule_symbols =
-            self.mangle_module_symbols(&submodule_symbols, submodule_path);
-        self.imported_symbols
-            .insert(namespace.to_string(), mangled_submodule_symbols);
-
-        let symbol = self.make_module_symbol(namespace, span);
-
-        resolver
-            .borrow_mut()
-            .cache_module(submodule_path, submodule_nodes.clone());
-        resolver.borrow_mut().finish_import(submodule_path);
-
-        self.all_module_asts
-            .insert(submodule_path.to_string(), submodule_nodes);
-
-        if !self
-            .module_dependencies
-            .contains(&submodule_path.to_string())
-        {
-            self.module_dependencies.push(submodule_path.to_string());
-        }
-
-        Ok(symbol)
-    }
-
-    fn handle_directory_import(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        resolver: std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut crate::diagnostic::Files,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-        let submodules = self.get_directory_submodules(module_path, span, &resolver)?;
-        match spec {
-            ImportSpec::Module { alias } => self.import_directory_as_module(
-                module_path,
-                alias.as_deref(),
-                &submodules,
-                span,
-                &resolver,
-                files,
-            )?,
-            ImportSpec::Item { item, alias } => {
-                self.ensure_directory_submodule_exists(module_path, item, &submodules, span)?;
-                self.import_directory_single_item(
-                    module_path,
-                    item,
-                    alias.as_deref(),
-                    span,
-                    &resolver,
-                    files,
-                )?
-            }
-            ImportSpec::Items { items } => self.import_directory_items(
-                module_path,
-                items,
-                &submodules,
-                span,
-                &resolver,
-                files,
-            )?,
-            ImportSpec::Wildcard => {
-                self.import_directory_wildcard(module_path, &submodules, span, &resolver, files)?
-            }
-        }
-        Ok(())
-    }
-
-    fn get_directory_submodules(
-        &self,
-        module_path: &str,
-        span: Span,
-        resolver: &std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-    ) -> Result<Vec<String>, SemanticError> {
-        resolver
-            .borrow()
-            .get_submodules(module_path)
-            .map_err(|e| SemanticError::new(format!("Failed to get submodules: {}", e), span))
-    }
-
-    fn ensure_directory_submodule_exists(
-        &self,
-        module_path: &str,
-        submodule_name: &str,
-        submodules: &[String],
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        if submodules.contains(&submodule_name.to_string()) {
-            return Ok(());
-        }
-        Err(SemanticError::with_help(
-            format!(
-                "Submodule '{}' not found in '{}'",
-                submodule_name, module_path
-            ),
-            span,
-            format!("Available submodules: {}", submodules.join(", ")),
-        ))
-    }
-
-    fn import_directory_as_module(
-        &mut self,
-        module_path: &str,
-        alias: Option<&str>,
-        submodules: &[String],
-        span: Span,
-        resolver: &std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut crate::diagnostic::Files,
-    ) -> Result<(), SemanticError> {
-        let namespace = alias.unwrap_or(module_path);
-        let mut module_symbols = std::collections::HashMap::new();
-        for submodule_name in submodules {
-            let submodule_path = format!("{}.{}", module_path, submodule_name);
-            let symbol = self.resolve_and_register_submodule(
-                &submodule_path,
-                submodule_name,
-                span,
-                resolver,
-                files,
-            )?;
-            module_symbols.insert(submodule_name.to_string(), symbol);
-        }
-        self.imported_symbols
-            .insert(namespace.to_string(), module_symbols);
-        self.symbol_table.add_symbol(
-            namespace,
-            Symbol {
-                kind: SymbolKind::Import,
-                span,
-                type_: Some(Type::Module(namespace.to_string())),
-                interfaces: std::collections::HashMap::new(),
-                methods: std::collections::HashMap::new(),
-                fields: std::collections::HashMap::new(),
-                type_params: Vec::new(),
-                original_name: None,
-                llvm_name: None,
-                default_param_count: 0,
-                variants: None,
-            },
-        )?;
-        Ok(())
-    }
-
-    fn import_directory_single_item(
-        &mut self,
-        module_path: &str,
-        item: &str,
-        alias: Option<&str>,
-        span: Span,
-        resolver: &std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut crate::diagnostic::Files,
-    ) -> Result<(), SemanticError> {
-        let submodule_path = format!("{}.{}", module_path, item);
-        let namespace = alias.unwrap_or(item);
-        let symbol =
-            self.resolve_and_register_submodule(&submodule_path, namespace, span, resolver, files)?;
-        self.symbol_table.add_symbol(namespace, symbol)?;
-        Ok(())
-    }
-
-    fn import_directory_items(
-        &mut self,
-        module_path: &str,
-        items: &[(String, Option<String>)],
-        submodules: &[String],
-        span: Span,
-        resolver: &std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut crate::diagnostic::Files,
-    ) -> Result<(), SemanticError> {
-        for (item, alias) in items {
-            self.ensure_directory_submodule_exists(module_path, item, submodules, span)?;
-            self.import_directory_single_item(
-                module_path,
-                item,
-                alias.as_deref(),
-                span,
-                resolver,
-                files,
-            )?;
-        }
-        Ok(())
-    }
-
-    fn import_directory_wildcard(
-        &mut self,
-        module_path: &str,
-        submodules: &[String],
-        span: Span,
-        resolver: &std::rc::Rc<std::cell::RefCell<crate::module_resolver::ModuleResolver>>,
-        files: &mut crate::diagnostic::Files,
-    ) -> Result<(), SemanticError> {
-        for submodule_name in submodules {
-            let submodule_path = format!("{}.{}", module_path, submodule_name);
-            let symbol = self.resolve_and_register_submodule(
-                &submodule_path,
-                submodule_name,
-                span,
-                resolver,
-                files,
-            )?;
-            self.symbol_table.add_symbol(submodule_name, symbol)?;
-        }
-        Ok(())
-    }
-
-    // Helper method to mangle module symbols
-    fn mangle_module_symbols(
-        &self,
-        symbols: &std::collections::HashMap<String, Symbol>,
-        module_path: &str,
-    ) -> std::collections::HashMap<String, Symbol> {
-        let module_name_for_mangling = Self::sanitize_module_path(module_path);
-        let mut mangled_symbols = std::collections::HashMap::new();
-
-        for (name, symbol) in symbols {
-            let mut mangled_symbol = symbol.clone();
-            if matches!(symbol.kind, SymbolKind::Function) {
-                mangled_symbol.llvm_name = Some(format!("{}!{}", module_name_for_mangling, name));
-            }
-            mangled_symbols.insert(name.clone(), mangled_symbol);
-        }
-
-        mangled_symbols
-    }
-
-    // Import single symbol (import logger.log)
-    fn import_single_symbol(
-        &mut self,
-        module_symbols: &std::collections::HashMap<String, Symbol>,
-        item_name: &str,
-        local_name: &str,
-        module_path: &str,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        let symbol = module_symbols.get(item_name).ok_or_else(|| {
-            let available: Vec<&String> = module_symbols.keys().collect();
-            if available.is_empty() {
-                SemanticError::new(
-                    format!(
-                        "Symbol '{}' not found in module '{}'",
-                        item_name, module_path
-                    ),
-                    span,
-                )
-            } else {
-                SemanticError::with_help(
-                    format!(
-                        "Symbol '{}' not found in module '{}'",
-                        item_name, module_path
-                    ),
-                    span,
-                    format!(
-                        "Available symbols: {}",
-                        available
-                            .iter()
-                            .map(|s| s.as_str())
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                )
-            }
-        })?;
-
-        // Clone the symbol and set original_name/llvm_name if needed
-        let mut imported_symbol = symbol.clone();
-
-        // Set original_name if it's an alias
-        if item_name != local_name {
-            imported_symbol.original_name = Some(item_name.to_string());
-        }
-
-        // Set llvm_name for functions (mangled with module path)
-        if matches!(symbol.kind, SymbolKind::Function) {
-            let module_name_for_mangling = module_path.replace(['.', '/'], "_");
-            imported_symbol.llvm_name = Some(format!("{}!{}", module_name_for_mangling, item_name));
-        }
-
-        self.add_import_symbol_if_absent(local_name, imported_symbol)?;
-        Ok(())
-    }
-
-    fn add_import_symbol_if_absent(
-        &mut self,
-        name: &str,
-        symbol: Symbol,
-    ) -> Result<(), SemanticError> {
-        if self.symbol_table.get_cloned(name).is_some() {
-            return Ok(());
-        }
-        self.symbol_table.add_imported_symbol(name, symbol)
-    }
-
-    // Import all symbols (import logger.*)
-    fn import_all_symbols(
-        &mut self,
-        module_symbols: &std::collections::HashMap<String, Symbol>,
-        module_path: &str,
-        _span: Span,
-    ) -> Result<(), SemanticError> {
-        let module_name_for_mangling = module_path.replace(['.', '/'], "_");
-
-        for (name, symbol) in module_symbols {
-            let mut imported_symbol = symbol.clone();
-
-            if matches!(symbol.kind, SymbolKind::Function) {
-                imported_symbol.llvm_name = Some(format!("{}!{}", module_name_for_mangling, name));
-            }
-
-            self.add_import_symbol_if_absent(name, imported_symbol)?;
-        }
-        Ok(())
-    }
-
-    /// Register a single built-in function into the symbol table from its signature.
-    fn register_builtin_function(&mut self, name: &str, sig: &BuiltInSig, span: Span) {
-        let _ = self.symbol_table.add_symbol(
-            name,
-            Self::make_symbol(
-                SymbolKind::Function,
-                span,
-                Some(Type::Function {
-                    params: sig.params.clone(),
-                    returns: Box::new(sig.return_type.clone()),
-                    default_count: 0,
-                }),
-            ),
-        );
-    }
-
-    /// Register all built-in functions whose names start with the given prefix.
-    fn register_builtin_functions_with_prefix(&mut self, prefix: &str, span: Span) {
-        let matching: Vec<_> = crate::semantics::stdlib::BUILT_IN_FUNCTIONS
-            .iter()
-            .filter(|(k, _)| k.starts_with(prefix))
-            .map(|(k, v)| (k.to_string(), v.clone()))
-            .collect();
-        for (func_name, sig) in matching {
-            self.register_builtin_function(&func_name, &sig, span);
-        }
-    }
-
-    fn import_std_item_with_resolver_check(
-        &mut self,
-        item: &str,
-        alias: &Option<String>,
-        span: Span,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-        let full_module_path = format!("std.{}", item);
-        if let Some(resolver) = &self.module_resolver {
-            let (has_file, has_directory) = resolver.borrow().check_module_path(&full_module_path);
-            if has_file || has_directory {
-                let files = files.ok_or_else(|| {
-                    SemanticError::new("Files registry must be available for std imports", span)
-                })?;
-                return self.import_module_from_resolver(
-                    &full_module_path,
-                    &ImportSpec::Module {
-                        alias: alias.clone(),
-                    },
-                    span,
-                    files,
-                );
-            }
-        }
-        self.import_stdlib_module(
-            item,
-            &ImportSpec::Module {
-                alias: alias.clone(),
-            },
-            span,
-        )
-    }
-
-    fn handle_std_import_with_std_prefix(
-        &mut self,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-        let mut files = files;
-        match spec {
-            ImportSpec::Item { item, alias } => {
-                self.import_std_item_with_resolver_check(item, alias, span, files.as_deref_mut())
-            }
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    self.import_std_item_with_resolver_check(
-                        item,
-                        alias,
-                        span,
-                        files.as_deref_mut(),
-                    )?;
-                }
-                Ok(())
-            }
-            _ => self.handle_std_import_all(spec, span),
-        }
-    }
-
-    fn files_for_std_import(
-        files: Option<&mut Files>,
-        span: Span,
-    ) -> Result<&mut Files, SemanticError> {
-        files.ok_or_else(|| SemanticError::new("Files registry must be available", span))
-    }
-
-    fn import_registry_std_module(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        files: Option<&mut Files>,
-        def: &crate::semantics::std_registry::StdModuleDef,
-    ) -> Result<(), SemanticError> {
-        match def.kind {
-            StdModuleKind::RuntimeBacked => {
-                let module_name = module_path.strip_prefix("std.").unwrap_or(module_path);
-                self.import_stdlib_module(module_name, spec, span)
-            }
-            StdModuleKind::Embedded => {
-                let files = Self::files_for_std_import(files, span)?;
-                self.import_module_from_resolver(module_path, spec, span, files)
-            }
-        }
-    }
-
-    fn import_nested_std_module_if_present(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        files: Option<&mut Files>,
-        registry: &std::collections::HashMap<
-            &'static str,
-            crate::semantics::std_registry::StdModuleDef,
-        >,
-    ) -> Result<Option<()>, SemanticError> {
-        let tail = match module_path.strip_prefix("std.") {
-            Some(tail) => tail,
-            None => return Ok(None),
-        };
-
-        let parent = match tail.split('.').next() {
-            Some(parent) => parent,
-            None => return Ok(None),
-        };
-
-        let parent_path = format!("std.{}", parent);
-        if !registry.contains_key(parent_path.as_str()) {
-            return Ok(None);
-        }
-
-        let def = match registry.get(module_path) {
-            Some(def) => def,
-            None => return Ok(None),
-        };
-
-        match def.kind {
-            StdModuleKind::RuntimeBacked => {
-                self.import_stdlib_module(tail, spec, span)?;
-            }
-            StdModuleKind::Embedded => {
-                let files = Self::files_for_std_import(files, span)?;
-                self.import_module_from_resolver(tail, spec, span, files)?;
-            }
-        }
-
-        Ok(Some(()))
-    }
-
-    fn handle_std_import(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        files: Option<&mut Files>,
-    ) -> Result<(), SemanticError> {
-        if module_path == "std" {
-            return self.handle_std_import_with_std_prefix(spec, span, files);
-        }
-
-        if let Some(resolver) = &self.module_resolver {
-            let (has_file, has_directory) = resolver.borrow().check_module_path(module_path);
-            if has_file || has_directory {
-                let files = Self::files_for_std_import(files, span)?;
-                return self.import_module_from_resolver(module_path, spec, span, files);
-            }
-        }
-
-        let registry = std_module_registry();
-        if let Some(def) = registry.get(module_path) {
-            return self.import_registry_std_module(module_path, spec, span, files, def);
-        }
-
-        if self
-            .import_nested_std_module_if_present(module_path, spec, span, files, registry)?
-            .is_some()
-        {
-            return Ok(());
-        }
-
-        match module_path {
-            "std" | "stdlib" => self.handle_std_import_all(spec, span),
-            _ => self.handle_non_stdlib_import(module_path, spec, span),
-        }
-    }
-
-    fn handle_std_import_all(
-        &mut self,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-        let registry = std_module_registry();
-
-        match spec {
-            ImportSpec::Module { alias } => {
-                self.import_all_std_as_namespace(alias.as_deref(), span, registry)?
-            }
-            ImportSpec::Wildcard => self.import_all_std_wildcard(span, registry)?,
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    self.import_stdlib_module(
-                        item,
-                        &ImportSpec::Module {
-                            alias: alias.clone(),
-                        },
-                        span,
-                    )?;
-                }
-            }
-            ImportSpec::Item { item, alias } => {
-                self.import_stdlib_module(
-                    item,
-                    &ImportSpec::Module {
-                        alias: alias.clone(),
-                    },
-                    span,
-                )?;
-            }
-        }
-        Ok(())
-    }
-
-    fn import_all_std_as_namespace(
-        &mut self,
-        alias: Option<&str>,
-        span: Span,
-        registry: &std::collections::HashMap<
-            &'static str,
-            crate::semantics::std_registry::StdModuleDef,
-        >,
-    ) -> Result<(), SemanticError> {
-        let namespace = alias.unwrap_or("std");
-        let namespace_symbols: std::collections::HashMap<String, Symbol> = registry
-            .keys()
-            .filter_map(|m| {
-                let rest = m.strip_prefix("std.")?;
-                if rest.contains('.') {
-                    None
-                } else {
-                    Some((rest.to_string(), self.make_module_symbol(m, span)))
-                }
-            })
-            .collect();
-
-        for (module_path, def) in registry {
-            for feature in def.runtime_features {
-                self.required_runtime_features
-                    .insert((*feature).to_string());
-            }
-
-            let short_name = module_path.strip_prefix("std.").unwrap_or(module_path);
-            if short_name.contains('.') {
-                continue;
-            }
-
-            let module_symbols = self.collect_stdlib_module_symbols(short_name, span);
-            self.imported_symbols
-                .insert(short_name.to_string(), module_symbols.clone());
-            self.imported_symbols
-                .insert(module_path.to_string(), module_symbols);
-            self.inject_nested_stdlib_children(short_name, span);
-        }
-
-        self.imported_symbols
-            .entry(namespace.to_string())
-            .or_insert(namespace_symbols);
-        self.add_import_symbol_if_absent(namespace, self.make_module_symbol(namespace, span))?;
-        Ok(())
-    }
-
-    fn import_all_std_wildcard(
-        &mut self,
-        span: Span,
-        registry: &std::collections::HashMap<
-            &'static str,
-            crate::semantics::std_registry::StdModuleDef,
-        >,
-    ) -> Result<(), SemanticError> {
-        for def in registry.values() {
-            for feature in def.runtime_features {
-                self.required_runtime_features
-                    .insert((*feature).to_string());
-            }
-        }
-
-        for (key, item) in crate::semantics::stdlib::all_stdlib_items() {
-            if let Some(item_name) = key.find('.').map(|i| &key[i + 1..]) {
-                let symbol = crate::semantics::stdlib::stdlib_item_to_symbol(&item, span);
-                self.add_import_symbol_if_absent(item_name, symbol)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn handle_non_stdlib_import(
-        &mut self,
-        module_path: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-
-        let module_name = module_path
-            .split('.')
-            .next_back()
-            .expect("module path should have at least one component");
-
-        match spec {
-            ImportSpec::Module { alias } => {
-                let symbol_name = alias.as_ref().map(|s| s.as_str()).unwrap_or(module_name);
-                if let Some(sig) = self.get_builtin_sig(symbol_name).cloned() {
-                    self.register_builtin_function(symbol_name, &sig, span);
-                } else if symbol_name == "none" {
-                    self.symbol_table.add_symbol(
-                        symbol_name,
-                        Self::make_symbol(
-                            SymbolKind::Constant,
-                            span,
-                            Some(Type::Optional(Box::new(Type::Void))),
-                        ),
-                    )?;
-                } else {
-                    self.register_builtin_functions_with_prefix(&format!("{}_", symbol_name), span);
-                }
-            }
-            ImportSpec::Item { item, alias } => {
-                let symbol_name = alias.as_ref().unwrap_or(item);
-                if let Some(sig) = self.get_builtin_sig(item).cloned() {
-                    self.register_builtin_function(symbol_name, &sig, span);
-                }
-            }
-            ImportSpec::Wildcard => {
-                self.register_builtin_functions_with_prefix(&format!("{}_", module_name), span);
-            }
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    let qualified_name = format!("{}_{}", module_name, item);
-                    let symbol_name = alias.as_ref().unwrap_or(&qualified_name);
-                    if let Some(sig) = self.get_builtin_sig(&qualified_name).cloned() {
-                        self.register_builtin_function(symbol_name, &sig, span);
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// Import a stdlib module (e.g., "math", "random")
-    fn import_stdlib_module(
-        &mut self,
-        module_name: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-    ) -> Result<(), SemanticError> {
-        self.track_runtime_features_for_std_module_name(module_name);
-        let module_symbols = self.collect_stdlib_module_symbols(module_name, span);
-        // Inject any nested stdlib children declared for this parent
-        self.inject_nested_stdlib_children(module_name, span);
-        self.apply_stdlib_module_import_spec(module_name, spec, span, module_symbols)
-    }
-
-    fn collect_stdlib_module_symbols(
-        &self,
-        module_name: &str,
-        span: Span,
-    ) -> std::collections::HashMap<String, Symbol> {
-        use crate::semantics::stdlib::all_stdlib_items;
-
-        let mut module_symbols = std::collections::HashMap::new();
-        let prefixes = Self::stdlib_module_prefixes(module_name);
-
-        for (key, item) in all_stdlib_items() {
-            if let Some(item_name) = Self::stdlib_item_name_for_module(&key, &prefixes) {
-                module_symbols.insert(
-                    item_name,
-                    crate::semantics::stdlib::stdlib_item_to_symbol(&item, span),
-                );
-            }
-        }
-
-        Self::inject_stdlib_module_special_symbols(module_name, span, &mut module_symbols);
-        module_symbols
-    }
-
-    fn stdlib_module_prefixes(module_name: &str) -> Vec<String> {
-        if !module_name.contains('.') {
-            return vec![module_name.to_string()];
-        }
-        let mut prefixes = vec![module_name.to_string()];
-        if let Some(last) = module_name.split('.').next_back() {
-            prefixes.push(last.to_string());
-        }
-        prefixes
-    }
-
-    fn stdlib_item_name_for_module(key: &str, prefixes: &[String]) -> Option<String> {
-        for prefix in prefixes {
-            let pattern = format!("{}.", prefix);
-            if let Some(rest) = key.strip_prefix(&pattern)
-                && !rest.contains('.')
-            {
-                return Some(rest.to_string());
-            }
-        }
-        None
-    }
-
-    fn inject_stdlib_module_special_symbols(
-        module_name: &str,
-        span: Span,
-        module_symbols: &mut std::collections::HashMap<String, Symbol>,
-    ) {
-        match module_name {
-            "net" => {
-                module_symbols.extend(crate::semantics::stdlib::net_module_class_symbols(span))
-            }
-            "sync" => {
-                module_symbols.extend(crate::semantics::stdlib::sync_module_class_symbols(span))
-            }
-            "sql" => {
-                module_symbols.extend(crate::semantics::stdlib::sql_module_class_symbols(span))
-            }
-            _ if module_name.ends_with(".json") => {
-                module_symbols.insert("Json".to_string(), Self::make_json_symbol(span));
-            }
-            _ if module_name.ends_with(".csv") => {
-                module_symbols.insert("Csv".to_string(), Self::make_csv_symbol(span));
-            }
-            _ => {}
-        }
-    }
-
-    fn csv_headers_type() -> Type {
-        Type::List(Box::new(Type::Primitive(PrimitiveType::Str)))
-    }
-
-    fn csv_rows_type() -> Type {
-        Type::List(Box::new(Self::csv_headers_type()))
-    }
-
-    fn csv_fields() -> std::collections::HashMap<String, (Type, bool)> {
-        let mut fields = std::collections::HashMap::new();
-        fields.insert("headers".to_string(), (Self::csv_headers_type(), true));
-        fields.insert("rows".to_string(), (Self::csv_rows_type(), true));
-        fields
-    }
-
-    fn make_csv_symbol(span: Span) -> Symbol {
-        let mut methods = std::collections::HashMap::new();
-        methods.insert(
-            "stringify".to_string(),
-            MethodSig {
-                params: vec![],
-                return_type: Type::Result(
-                    Box::new(Type::Primitive(PrimitiveType::Str)),
-                    Box::new(Type::Primitive(PrimitiveType::Str)),
-                ),
-                is_static: false,
-            },
-        );
-        Symbol {
-            kind: SymbolKind::Class,
-            span,
-            type_: Some(Type::Named("Csv".to_string(), Vec::new())),
-            interfaces: std::collections::HashMap::new(),
-            methods,
-            fields: Self::csv_fields(),
-            type_params: Vec::new(),
-            original_name: None,
-            llvm_name: None,
-            default_param_count: 0,
-            variants: None,
-        }
-    }
-
-    fn make_json_symbol(span: Span) -> Symbol {
-        let mut methods = std::collections::HashMap::new();
-        methods.insert(
-            "stringify".to_string(),
-            MethodSig {
-                params: vec![Type::Optional(Box::new(Type::Primitive(
-                    PrimitiveType::Int,
-                )))],
-                return_type: Type::Result(
-                    Box::new(Type::Primitive(PrimitiveType::Str)),
-                    Box::new(Type::Primitive(PrimitiveType::Str)),
-                ),
-                is_static: false,
-            },
-        );
-        Symbol {
-            kind: SymbolKind::Class,
-            span,
-            type_: Some(Type::Named("Json".to_string(), Vec::new())),
-            interfaces: std::collections::HashMap::new(),
-            methods,
-            fields: std::collections::HashMap::new(),
-            type_params: Vec::new(),
-            original_name: None,
-            llvm_name: None,
-            default_param_count: 0,
-            variants: None,
-        }
-    }
-
-    fn apply_stdlib_module_import_spec(
-        &mut self,
-        module_name: &str,
-        spec: &crate::ast::ImportSpec,
-        span: Span,
-        module_symbols: std::collections::HashMap<String, Symbol>,
-    ) -> Result<(), SemanticError> {
-        use crate::ast::ImportSpec;
-
-        match spec {
-            ImportSpec::Module { alias } => {
-                let namespace = alias.as_deref().map(|s| s.to_string()).unwrap_or_else(|| {
-                    // If importing a nested module like "data.json" prefer the
-                    // short child name "json" as the namespace so callers can
-                    // reference `json.parse` without an alias.
-                    if module_name.contains('.') {
-                        module_name.split('.').next_back().unwrap().to_string()
-                    } else {
-                        module_name.to_string()
-                    }
-                });
-
-                self.imported_symbols
-                    .entry(namespace.to_string())
-                    .or_insert(module_symbols.clone());
-                self.add_import_symbol_if_absent(
-                    &namespace,
-                    Symbol {
-                        kind: SymbolKind::Import,
-                        span,
-                        type_: Some(Type::Module(namespace.to_string())),
-                        interfaces: std::collections::HashMap::new(),
-                        methods: std::collections::HashMap::new(),
-                        fields: std::collections::HashMap::new(),
-                        type_params: Vec::new(),
-                        original_name: None,
-                        llvm_name: None,
-                        default_param_count: 0,
-                        variants: None,
-                    },
-                )?;
-            }
-            ImportSpec::Item { item, alias } => {
-                let symbol_name = alias.as_ref().unwrap_or(item);
-                if let Some(symbol) = module_symbols.get(item) {
-                    self.add_import_symbol_if_absent(symbol_name, symbol.clone())?;
-                }
-            }
-            ImportSpec::Wildcard => {
-                for (name, symbol) in module_symbols {
-                    self.add_import_symbol_if_absent(&name, symbol)?;
-                }
-            }
-            ImportSpec::Items { items } => {
-                for (item, alias) in items {
-                    let symbol_name = alias.as_ref().unwrap_or(item);
-                    if let Some(symbol) = module_symbols.get(item) {
-                        self.add_import_symbol_if_absent(symbol_name, symbol.clone())?;
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    // stdlib item conversion/registration delegated to `crate::semantics::stdlib`.
-
-    // Helper to find free variables in a block of statements
-    // Returns variables that are used but not declared in the local scope
-    fn find_free_variables_in_block(
-        &self,
-        body: &[StatementNode],
-        local_vars: &std::collections::HashSet<String>,
-    ) -> Result<Vec<(String, Type)>, SemanticError> {
-        let mut free_vars = std::collections::HashMap::new();
-        let mut local_vars = local_vars.clone();
-
-        for stmt in body {
-            self.find_free_variables_in_statement(stmt, &mut local_vars, &mut free_vars)?;
-        }
-
-        Ok(free_vars.into_iter().collect())
-    }
-
-    fn find_free_variables_in_statement(
-        &self,
-        stmt: &StatementNode,
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        match &stmt.kind {
-            StatementKind::Expression(expr) | StatementKind::Return(Some(expr)) => {
-                self.find_free_variables_in_expression(expr, local_vars, free_vars)?;
-            }
-            StatementKind::AutoDecl(name, _, expr) => {
-                self.handle_variable_declaration(name, expr, local_vars, free_vars)?;
-            }
-            StatementKind::TypedDecl(name, _, expr) => {
-                self.handle_variable_declaration(name, expr, local_vars, free_vars)?;
-            }
-            StatementKind::ConstDecl(name, _, expr) => {
-                self.handle_variable_declaration(name, expr, local_vars, free_vars)?;
-            }
-            StatementKind::If {
-                cond,
-                then_block,
-                else_block,
-            } => {
-                self.handle_if_statement(cond, then_block, else_block, local_vars, free_vars)?;
-            }
-            StatementKind::While { cond, body } => {
-                self.handle_while_statement(cond, body, local_vars, free_vars)?;
-            }
-            StatementKind::For {
-                var, iter, body, ..
-            } => {
-                self.handle_for_statement(var, iter, body, local_vars, free_vars)?;
-            }
-            StatementKind::Block(stmts) => {
-                self.handle_block_statement(stmts, local_vars, free_vars)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn handle_variable_declaration(
-        &self,
-        name: &str,
-        expr: &ExpressionNode,
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        // First analyze the expression (uses happen before the decl is in scope)
-        self.find_free_variables_in_expression(expr, local_vars, free_vars)?;
-        // Then add the variable to local scope
-        local_vars.insert(name.to_string());
-        Ok(())
-    }
-
-    fn handle_if_statement(
-        &self,
-        cond: &ExpressionNode,
-        then_block: &[StatementNode],
-        else_block: &Option<Vec<StatementNode>>,
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(cond, local_vars, free_vars)?;
-        for s in then_block {
-            self.find_free_variables_in_statement(s, local_vars, free_vars)?;
-        }
-        if let Some(else_stmts) = else_block {
-            for s in else_stmts {
-                self.find_free_variables_in_statement(s, local_vars, free_vars)?;
-            }
-        }
-        Ok(())
-    }
-
-    fn handle_while_statement(
-        &self,
-        cond: &ExpressionNode,
-        body: &[StatementNode],
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(cond, local_vars, free_vars)?;
-        for s in body {
-            self.find_free_variables_in_statement(s, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_for_statement(
-        &self,
-        var: &str,
-        iter: &ExpressionNode,
-        body: &[StatementNode],
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(iter, local_vars, free_vars)?;
-        // Iterator variable is local to the for loop
-        local_vars.insert(var.to_string());
-        for s in body {
-            self.find_free_variables_in_statement(s, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_block_statement(
-        &self,
-        stmts: &[StatementNode],
-        local_vars: &mut std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        for s in stmts {
-            self.find_free_variables_in_statement(s, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn find_free_variables_in_expression(
-        &self,
-        expr: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        match &expr.kind {
-            ExpressionKind::Identifier(name) => {
-                self.handle_identifier_expression(name, local_vars, free_vars)?;
-            }
-            ExpressionKind::Binary { left, right, .. } => {
-                self.handle_binary_expression(left, right, local_vars, free_vars)?;
-            }
-            ExpressionKind::Unary {
-                expr: inner,
-                op_span: _,
-                ..
-            } => {
-                self.handle_unary_expression(inner, local_vars, free_vars)?;
-            }
-            ExpressionKind::Call { func, args } => {
-                self.handle_call_expression(func, args, local_vars, free_vars)?;
-            }
-            ExpressionKind::FieldAccess { expr: inner, .. } => {
-                self.handle_field_access_expression(inner, local_vars, free_vars)?;
-            }
-            ExpressionKind::ListAccess { expr: inner, index } => {
-                self.handle_list_access_expression(inner, index, local_vars, free_vars)?;
-            }
-            ExpressionKind::ListLiteral(elements) => {
-                self.handle_list_literal_expression(elements, local_vars, free_vars)?;
-            }
-            ExpressionKind::MapLiteral { entries, .. } => {
-                self.handle_map_literal_expression(entries, local_vars, free_vars)?;
-            }
-            ExpressionKind::SetOrMapLiteral(elements) => {
-                self.handle_set_literal_expression(elements, local_vars, free_vars)?;
-            }
-            ExpressionKind::TupleLiteral(elements) => {
-                self.handle_tuple_literal_expression(elements, local_vars, free_vars)?;
-            }
-            ExpressionKind::If {
-                cond,
-                then_expr,
-                else_expr,
-            } => {
-                self.handle_if_expression(cond, then_expr, else_expr, local_vars, free_vars)?;
-            }
-            ExpressionKind::Lambda { params, body, .. } => {
-                self.handle_lambda_expression(params, body, local_vars, free_vars)?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    fn handle_identifier_expression(
-        &self,
-        name: &str,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        // Check if it's a local variable (parameter or declared in lambda body)
-        if !local_vars.contains(name) {
-            // Check if it exists in outer scopes
-            if let Some(symbol) = self.symbol_table.lookup(name)
-                && matches!(symbol.kind, SymbolKind::Variable)
-                && let Some(var_type) = &symbol.type_
-            {
-                free_vars.insert(name.to_string(), var_type.clone());
-            }
-        }
-        Ok(())
-    }
-
-    fn handle_binary_expression(
-        &self,
-        left: &ExpressionNode,
-        right: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(left, local_vars, free_vars)?;
-        self.find_free_variables_in_expression(right, local_vars, free_vars)?;
-        Ok(())
-    }
-
-    fn handle_unary_expression(
-        &self,
-        inner: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(inner, local_vars, free_vars)?;
-        Ok(())
-    }
-
-    fn handle_call_expression(
-        &self,
-        func: &ExpressionNode,
-        args: &[ExpressionNode],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(func, local_vars, free_vars)?;
-        for arg in args {
-            self.find_free_variables_in_expression(arg, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_field_access_expression(
-        &self,
-        inner: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(inner, local_vars, free_vars)?;
-        Ok(())
-    }
-
-    fn handle_list_access_expression(
-        &self,
-        inner: &ExpressionNode,
-        index: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(inner, local_vars, free_vars)?;
-        self.find_free_variables_in_expression(index, local_vars, free_vars)?;
-        Ok(())
-    }
-
-    fn handle_list_literal_expression(
-        &self,
-        elements: &[ExpressionNode],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        for elem in elements {
-            self.find_free_variables_in_expression(elem, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_map_literal_expression(
-        &self,
-        entries: &[(ExpressionNode, ExpressionNode)],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        for (key, val) in entries {
-            self.find_free_variables_in_expression(key, local_vars, free_vars)?;
-            self.find_free_variables_in_expression(val, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_set_literal_expression(
-        &self,
-        elements: &[ExpressionNode],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        for elem in elements {
-            self.find_free_variables_in_expression(elem, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_tuple_literal_expression(
-        &self,
-        elements: &[ExpressionNode],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        for elem in elements {
-            self.find_free_variables_in_expression(elem, local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn handle_if_expression(
-        &self,
-        cond: &ExpressionNode,
-        then_expr: &ExpressionNode,
-        else_expr: &ExpressionNode,
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        self.find_free_variables_in_expression(cond, local_vars, free_vars)?;
-        self.find_free_variables_in_expression(then_expr, local_vars, free_vars)?;
-        self.find_free_variables_in_expression(else_expr, local_vars, free_vars)?;
-        Ok(())
-    }
-
-    fn handle_lambda_expression(
-        &self,
-        params: &[Param],
-        body: &[StatementNode],
-        local_vars: &std::collections::HashSet<String>,
-        free_vars: &mut std::collections::HashMap<String, Type>,
-    ) -> Result<(), SemanticError> {
-        let mut inner_local_vars = local_vars.clone();
-        for param in params {
-            inner_local_vars.insert(param.name.clone());
-        }
-        for stmt in body {
-            self.find_free_variables_in_statement(stmt, &mut inner_local_vars, free_vars)?;
-        }
-        Ok(())
-    }
-
-    fn check_not_modifying_constant(
-        &mut self,
-        expr: &ExpressionNode,
-        op_span: &Span,
-    ) -> Result<(), SemanticError> {
-        if let crate::ast::ExpressionKind::Identifier(name) = &expr.kind {
-            if let Some(symbol) = self.symbol_table.lookup(name)
-                && symbol.kind == SymbolKind::Constant
-            {
-                return Err(SemanticError::with_help(
-                    format!("Cannot modify constant '{}'", name),
-                    *op_span,
-                    "Constants cannot be modified after initialization",
-                ));
-            }
-        } else if let crate::ast::ExpressionKind::FieldAccess {
-            expr: obj_expr,
-            field,
-        } = &expr.kind
-        {
-            let obj_type = self.get_expression_type(obj_expr)?;
-            if let Type::Named(class_name, _) = &obj_type
-                && let Some(symbol) = self.symbol_table.lookup(class_name)
-                && let Some((_, is_const)) = symbol.fields.get(field)
-                && *is_const
-            {
-                return Err(SemanticError::with_help(
-                    format!("Cannot modify const field '{}'", field),
-                    *op_span,
-                    "Const fields cannot be modified after initialization. Remove the 'const' modifier from the field declaration if mutation is needed.",
-                ));
-            }
-        }
-        Ok(())
     }
 }
